fixes for selection and details pane navigation, js and html separated into files, restructure of project files
