+++ src/bargePanel.ts
@@ -1,4 +1,7 @@
 import * as vscode from 'vscode';
+import * as fs from 'fs';
+import * as path from 'path';
+import * as crypto from 'crypto';
 import { AzureService } from './azureService';
 import { WebviewMessage, QueryResponse } from './types';
 
@@ -21,7 +24,7 @@ export class BargePanel {
         // Create panel in bottom area by default
         const panel = vscode.window.createWebviewPanel(
             BargePanel.viewType,
-            'bARGE - basic Azure Resource Graph Explorer',
+            'bARGE - boosted Azure Resource Graph Explorer',
             { viewColumn: vscode.ViewColumn.Active, preserveFocus: false },
             {
                 enableScripts: true,
@@ -31,7 +34,7 @@ export class BargePanel {
         );
 
         BargePanel.currentPanel = new BargePanel(panel, extensionUri, azureService);
-        
+
         // Move to bottom panel after creation
         vscode.commands.executeCommand('workbench.action.moveEditorToBelowGroup');
     }
@@ -70,34 +73,34 @@ export class BargePanel {
             this._panel.webview.postMessage({
                 type: 'queryStart'
             });
-            
+
             this._panel.reveal();
-            
+
             const result = await this._azureService.runQuery(query);
-            
+
             if (result && result.columns) {
                 result.columns = result.columns.map((col: any) => ({
                     ...col,
                     name: col.name.split(' (')[0]
                 }));
             }
-            
+
             const response: QueryResponse = {
                 success: true,
                 data: result
             };
-            
+
             this._panel.webview.postMessage({
                 type: 'queryResult',
                 payload: response
             });
-            
+
         } catch (error) {
             const response: QueryResponse = {
                 success: false,
                 error: String(error)
             };
-            
+
             this._panel.webview.postMessage({
                 type: 'queryResult',
                 payload: response
@@ -108,7 +111,7 @@ export class BargePanel {
     private async _exportToCsv(queryResult: any, filename: string) {
         try {
             const csvContent = this._convertToCsv(queryResult);
-            
+
             const uri = await vscode.window.showSaveDialog({
                 defaultUri: vscode.Uri.file(filename),
                 filters: {
@@ -132,10 +135,10 @@ export class BargePanel {
         }
 
         const headers = queryResult.columns.map((col: any) => col.name).join(',');
-        const rows = queryResult.data.map((row: any[]) => 
-            row.map(cell => 
+        const rows = queryResult.data.map((row: any[]) =>
+            row.map(cell =>
                 typeof cell === 'string' && (cell.includes(',') || cell.includes('"') || cell.includes('\n'))
-                    ? `"${cell.replace(/"/g, '""')}"` 
+                    ? `"${cell.replace(/"/g, '""')}"`
                     : String(cell || '')
             ).join(',')
         );
@@ -159,3039 +162,26 @@ export class BargePanel {
     }
 
     private _getHtmlForWebview() {
-        const loadingGifUri = this._panel.webview.asWebviewUri(
-            vscode.Uri.joinPath(this._extensionUri, 'media', 'loadingbarge.gif')
+        // Generate a nonce for Content Security Policy
+        const nonce = crypto.randomBytes(16).toString('base64');
+
+        const htmlPath = path.join(this._extensionUri.fsPath, 'media', 'webview', 'webview.html');
+        let html = fs.readFileSync(htmlPath, 'utf8');
+
+        // Get URIs for resources
+        const webviewUri = this._panel.webview.asWebviewUri(
+            vscode.Uri.joinPath(this._extensionUri, 'media', 'webview')
         );
 
-        return `<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>bARGE - basic Azure Resource Graph Explorer</title>
-    <style>
-        body {
-            font-family: var(--vscode-font-family);
-            font-size: var(--vscode-font-size);
-            color: var(--vscode-foreground);
-            background-color: var(--vscode-editor-background);
-            margin: 0;
-            padding: 0;
-            height: 100vh;
-            display: flex;
-            flex-direction: column;
-            overflow: hidden;
-        }
-        .main-content {
-            flex: 1;
-            display: flex;
-            flex-direction: column;
-            overflow: hidden;
-            position: relative;
-        }
-        .content-wrapper {
-            flex: 1;
-            display: flex;
-            position: relative;
-            overflow: hidden;
-        }
-        
-        /* Adaptive layout classes */
-        .content-wrapper.layout-horizontal {
-            flex-direction: row;
-        }
-        .content-wrapper.layout-vertical {
-            flex-direction: column;
-        }
-        
-        /* Table section */
-        .table-section {
-            display: flex;
-            flex-direction: column;
-            overflow: hidden;
-        }
-        .table-container {
-            flex: 1;
-            overflow: auto;
-            margin: 15px;
-            border: 1px solid var(--vscode-widget-border);
-            border-radius: 4px;
-            min-height: 200px;
-        }
-        
-        /* Details section */
-        .details-section {
-            display: flex;
-            flex-direction: column;
-            overflow: hidden;
-        }
-        .details-panel {
-            flex: 1;
-            background-color: var(--vscode-sideBar-background);
-            border: 1px solid var(--vscode-widget-border);
-            border-radius: 4px;
-            display: flex;
-            flex-direction: column;
-            overflow: hidden;
-            margin: 15px;
-        }
-        
-        /* Horizontal layout (side-by-side) */
-        .content-wrapper.layout-horizontal .table-section {
-            flex: 2;
-            min-width: 300px;
-        }
-        .content-wrapper.layout-horizontal .table-container {
-            margin-right: 8px;
-        }
-        .content-wrapper.layout-horizontal .details-section {
-            flex: 1;
-            min-width: 250px;
-        }
-        .content-wrapper.layout-horizontal .details-panel {
-            margin-left: 8px;
-        }
-        
-        /* Vertical layout (stacked) */
-        .content-wrapper.layout-vertical .table-section {
-            flex: 2;
-            min-height: 200px;
-        }
-        .content-wrapper.layout-vertical .table-container {
-            margin-bottom: 8px;
-        }
-        .content-wrapper.layout-vertical .details-section {
-            flex: 1;
-            min-height: 150px;
-        }
-        .content-wrapper.layout-vertical .details-panel {
-            margin-top: 8px;
-        }
-        
-        /* Resize handle - positioned absolutely between sections */
-        #resizeHandle.resize-handle {
-            position: absolute;
-            background-color: var(--vscode-panel-border, var(--vscode-widget-border, #3c3c3c));
-            transition: background-color 0.2s ease, opacity 0.2s ease;
-            z-index: 100;
-            opacity: 0.8;
-            border-radius: 3px;
-        }
-        #resizeHandle.resize-handle:hover {
-            background-color: var(--vscode-focusBorder, #007ACC);
-            opacity: 1;
-        }
-        #resizeHandle.resize-handle:active {
-            background-color: var(--vscode-focusBorder, #007ACC);
-            opacity: 1;
-        }
-        #resizeHandle.resize-handle.dragging {
-            transition: none; /* Remove all transitions while dragging */
-        }
-        
-        /* Horizontal layout resize handle (vertical divider) */
-        .content-wrapper.layout-horizontal #resizeHandle.resize-handle {
-            width: 6px;
-            height: 95%;
-            top: 2.5%;
-            cursor: ew-resize;
-        }
-        
-        /* Vertical layout resize handle (horizontal divider) */
-        .content-wrapper.layout-vertical #resizeHandle.resize-handle {
-            width: 95%;
-            height: 6px;
-            cursor: ns-resize;
-        }
-        .resize-handle:hover {
-            background-color: var(--vscode-focusBorder);
-        }
-        
-        .details-panel.hidden {
-            transform: translateX(100%);
-        }
-        .main-content.vertical-layout .details-panel.hidden {
-            transform: translateY(100%);
-        }
-        .details-panel.hidden {
-            transform: translateX(100%);
-        }
-        .details-header {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            padding: 12px 16px;
-            border-bottom: 1px solid var(--vscode-widget-border);
-            background-color: var(--vscode-editor-widget-background);
-            flex-shrink: 0;
-        }
-        .details-title {
-            font-weight: 600;
-            font-size: 1.1em;
-            color: var(--vscode-foreground);
-        }
-        .details-navigation {
-            display: flex;
-            gap: 4px;
-            align-items: center;
-        }
-        .nav-btn, .close-btn {
-            background-color: var(--vscode-button-background);
-            color: var(--vscode-button-foreground);
-            border: none;
-            width: 32px;
-            height: 32px;
-            border-radius: 4px;
-            cursor: pointer;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            font-size: 16px;
-            transition: background-color 0.2s ease;
-        }
-        .nav-btn:hover:not(:disabled), .close-btn:hover {
-            background-color: var(--vscode-button-hoverBackground);
-        }
-        .nav-btn:active:not(:disabled), .close-btn:active {  
-            background-color: var(--vscode-button-hoverBackground);  
-            transform: scale(0.95);  
-            transition: transform 0.1s ease;  
-        }
-        .nav-btn:disabled {
-            background-color: var(--vscode-button-secondaryBackground);
-            color: var(--vscode-disabledForeground);
-            cursor: not-allowed;
-        }
-        .details-content {
-            flex: 1;
-            overflow: auto;
-            padding: 16px;
-            min-width: 0; /* Allow shrinking below content width */
-            word-wrap: break-word; /* Break long words */
-            white-space: pre-wrap; /* Preserve formatting but allow wrapping */
-        }
-        /* Reduce padding when showing comparison table */
-        .details-content.comparison-view {
-            padding: 8px;
-        }
-        .detail-button {
-            background: transparent;
-            color: var(--vscode-descriptionForeground);
-            border: none;
-            width: 28px;
-            height: 28px;
-            border-radius: 3px;
-            cursor: pointer;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            font-size: 12px;
-            transition: all 0.2s ease;
-            opacity: 0.6;
-            margin: auto;
-            /* Prevent focus outline */
-            outline: none !important;
-        }
-        .detail-button:hover {
-            color: var(--vscode-foreground);
-            opacity: 1;
-            transform: scale(1.1);
-        }
-        .detail-button:active {
-            background-color: var(--vscode-list-activeSelectionBackground);
-            transform: scale(0.95);
-        }
-        .detail-button:focus,
-        .detail-button:focus-visible {
-            outline: none !important;
-            box-shadow: none !important;
-            background-color: transparent !important;
-        }
-        .detail-button.active {
-            color: var(--vscode-focusBorder);
-            opacity: 1;
-        }
-        .detail-button svg {
-            width: 24px;
-            height: 24px;
-            fill: currentColor;
-        }
-        .detail-button-cell {
-            width: 36px !important;
-            min-width: 36px !important;
-            max-width: 36px !important;
-            text-align: center;
-            padding: 2px !important;
-            vertical-align: middle;
-            position: relative;
-            /* Prevent focus outline issues */
-            outline: none !important;
-        }
-        .detail-button-cell:focus,
-        .detail-button-cell:focus-visible {
-            outline: none !important;
-            box-shadow: none !important;
-        }
-        /* Override any table cell focus styles specifically for detail button cells */
-        .results-table .detail-button-cell:focus,
-        .results-table .detail-button-cell:focus-visible,
-        .results-table td.detail-button-cell:focus,
-        .results-table td.detail-button-cell:focus-visible {
-            outline: none !important;
-            box-shadow: none !important;
-            background-color: transparent !important;
-        }
-        th.detail-button-cell {
-            cursor: default !important;
-            user-select: none !important;
-            position: relative !important;
-            pointer-events: auto !important;
-        }
-        /* Make header magnifying glass clickable with hover feedback */
-        th.detail-button-cell.select-all-header {
-            cursor: pointer !important;
-            transition: all 0.2s ease;
-        }
-        th.detail-button-cell.select-all-header:hover {
-            background-color: var(--vscode-list-hoverBackground) !important;
-        }
-        th.detail-button-cell.select-all-header:hover svg {
-            opacity: 1 !important;
-            transform: scale(1.1);
-        }
-        th.detail-button-cell.select-all-header:active {
-            background-color: var(--vscode-list-activeSelectionBackground) !important;
-            transform: scale(0.95);
-        }
-        th.detail-button-cell::after {
-            display: none !important;
-        }
-        th.detail-button-cell::before {
-            display: none !important;
-        }
-        th.detail-button-cell .resize-handle {
-            display: none !important;
-        }
-        th.detail-button-cell:hover::after {
-            display: none !important;
-        }
-        th.detail-button-cell * {
-            pointer-events: none !important;
-        }
-        th.detail-button-cell {
-            overflow: hidden !important;
-        }
-        .detail-button-cell:hover {
-            background-color: transparent !important;
-        }
-        /* Additional comprehensive focus prevention for detail buttons */
-        .detail-button-cell *,
-        .detail-button-cell button,
-        .detail-button-cell svg {
-            outline: none !important;
-            box-shadow: none !important;
-        }
-        .detail-button-cell *:focus,
-        .detail-button-cell *:focus-visible,
-        .detail-button-cell button:focus,
-        .detail-button-cell button:focus-visible {
-            outline: none !important;
-            box-shadow: none !important;
-            background-color: transparent !important;
-        }
-        .json-viewer {
-            font-family: var(--vscode-editor-font-family, 'Consolas', 'Monaco', monospace);
-            font-size: 0.9em;
-            line-height: 1.4;
-            color: var(--vscode-editor-foreground);
-            padding: 4px;
-        }
-        .json-property {
-            margin: 0;
-            padding: 8px 4px;
-            border-bottom: 1px solid var(--vscode-widget-border);
-            transition: background-color 0.2s ease;
-        }
-        .json-property:last-child {
-            border-bottom: none;
-        }
-        .json-property:hover {
-            background-color: rgba(255, 255, 255, 0.03);
-        }
-        .json-key {
-            color: var(--vscode-symbolIcon-propertyForeground, #9cdcfe);
-            font-weight: 600;
-            font-size: 0.85em;
-            text-transform: uppercase;
-            letter-spacing: 0.5px;
-            margin-bottom: 6px;
-            display: block;
-            opacity: 0.8;
-        }
-        .json-value {
-            color: var(--vscode-editor-foreground);
-            font-size: 0.95em;
-            line-height: 1.3;
-            word-wrap: break-word;
-            display: block;
-        }
-        .json-value.string {
-            color: var(--vscode-symbolIcon-stringForeground, #ce9178);
-            font-style: normal;
-        }
-        .json-value.number {
-            color: var(--vscode-symbolIcon-numberForeground, #b5cea8);
-            font-weight: 500;
-        }
-        .json-value.boolean {
-            color: var(--vscode-symbolIcon-booleanForeground, #569cd6);
-            font-weight: 500;
-        }
-        .json-value.null {
-            color: var(--vscode-symbolIcon-nullForeground, #808080);
-            font-style: italic;
-        }
-        .json-value.object {
-            color: var(--vscode-editor-foreground);
-            background-color: var(--vscode-textCodeBlock-background);
-            padding: 8px;
-            border-radius: 4px;
-            font-family: var(--vscode-editor-font-family, 'Consolas', 'Monaco', monospace);
-            font-size: 0.85em;
-            border-left: 3px solid var(--vscode-widget-border);
-            margin-top: 4px;
-        }
-        
-        /* Comparison view styles */
-        .comparison-viewer {
-            font-family: var(--vscode-editor-font-family, 'Consolas', 'Monaco', monospace);
-            font-size: 0.9em;
-            line-height: 1.4;
-            color: var(--vscode-editor-foreground);
-            padding: 4px;
-        }
-        .comparison-table {
-            width: 100%;
-            border-collapse: collapse;
-            border: 1px solid var(--vscode-widget-border);
-        }
-        .comparison-table th,
-        .comparison-table td {
-            padding: 8px;
-            border: 1px solid var(--vscode-widget-border);
-            text-align: left;
-            vertical-align: top;
-        }
-        .comparison-table th {
-            background-color: var(--vscode-list-hoverBackground);
-            font-weight: 600;
-            position: relative;
-            /* Removed sticky positioning */
-        }
-        .comparison-table th:not(:last-child) {
-            border-right: 1px solid var(--vscode-widget-border);
-        }
-        .comparison-table th .column-resizer {
-            position: absolute;
-            top: 0;
-            right: 0;
-            width: 4px;
-            height: 100%;
-            cursor: col-resize;
-            background: transparent;
-            z-index: 1;
-        }
-        .comparison-table th .column-resizer:hover {
-            background-color: var(--vscode-focusBorder);
-        }
-        .property-name-header {
-            width: auto;
-            min-width: 160px; /* Increased from 120px to accommodate text + indicator */
-            max-width: none;
-        }
-        .row-header {
-            min-width: 150px;
-        }
-        .property-name-with-indicator {
-            font-weight: 500;
-            color: var(--vscode-symbolIcon-propertyForeground, #9cdcfe);
-            word-break: break-word;
-            position: relative;
-            /* Remove flex layout */
-            min-height: calc(100% - 16px);
-        }
-        .property-name-text {
-            /* Add right margin to make space for the indicator */
-            margin-right: 24px; /* Increased from 20px to give more space */
-        }
-        .match-indicator-inline {
-            /* Position the indicator to align with text baseline */
-            position: absolute;
-            top: 8px; /* Match the cell's top padding to align with text */
-            right: 4px; /* Small padding from the right edge */
-            display: inline-block;
-            line-height: 1; /* Ensure consistent height */
-        }
-        .property-value {
-            word-break: break-word;
-            max-width: 300px;
-        }
-        .null-value {
-            color: var(--vscode-symbolIcon-nullForeground, #808080);
-            font-style: italic;
-        }
-        .object-value {
-            background-color: var(--vscode-textCodeBlock-background);
-            padding: 4px;
-            border-radius: 4px;
-            font-size: 0.85em;
-            margin: 0;
-            max-height: 200px;
-            overflow: auto;
-        }
-        .scalar-value {
-            color: var(--vscode-editor-foreground);
-        }
-        .property-row:hover {
-            background-color: rgba(255, 255, 255, 0.02);
-        }
-        /* Property column styling to fix border issues */
-        .comparison-table .property-row td:first-child {
-            border-right: 2px solid var(--vscode-widget-border);
-            background-color: rgba(255, 255, 255, 0.02);
-            position: relative;
-            /* Different approach: force the cell to use table-cell display with proper height */
-            height: 100%;
-        }
-        .comparison-table .property-row td:first-child::before {
-            content: '';
-            position: absolute;
-            top: 0;
-            left: 0;
-            right: 0;
-            bottom: 0;
-            background-color: inherit;
-            z-index: -1;
-        }
-        /* Ensure proper vertical alignment for property cells */
-        .comparison-table .property-row td {
-            vertical-align: top;
-        }
-        
-        .footer {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            padding: 8px 15px;
-            border-top: 1px solid var(--vscode-widget-border);
-            background-color: var(--vscode-editor-background);
-            flex-shrink: 0;
-            min-height: 30px;
-        }
-        .results-info {
-            font-size: 0.9em;
-            color: var(--vscode-descriptionForeground);
-        }
-        .export-btn {
-            background-color: var(--vscode-button-background);
-            color: var(--vscode-button-foreground);
-            border: none;
-            padding: 6px 12px;
-            border-radius: 4px;
-            cursor: pointer;
-            font-size: 0.9em;
-        }
-        .export-btn:hover {
-            background-color: var(--vscode-button-hoverBackground);
-        }
-        .results-table {
-            width: 100%;
-            border-collapse: collapse;
-            font-size: 0.9em;
-            table-layout: fixed;
-        }
-        .results-table th {
-            background-color: var(--vscode-sideBar-background, var(--vscode-editor-background));
-            border: 1px solid var(--vscode-widget-border);
-            padding: 8px;
-            text-align: left;
-            position: sticky;
-            top: -1px;
-            cursor: pointer;
-            user-select: none;
-            z-index: 10;
-            min-width: 60px;
-            resize: horizontal;
-            overflow: hidden;
-        }
-        .results-table th:hover {
-            background-color: var(--vscode-list-hoverBackground);
-        }
-        .results-table th.sorted-asc::after {
-            content: ' ↑';
-        }
-        .results-table th.sorted-desc::after {
-            content: ' ↓';
-        }
-        .results-table th.dragging {
-            opacity: 0.5;
-            background-color: var(--vscode-list-activeSelectionBackground);
-        }
-        .results-table th.drag-over {
-            border-left: 3px solid var(--vscode-focusBorder);
-        }
-        .resize-handle {
-            position: absolute;
-            top: 0;
-            right: 0;
-            width: 5px;
-            height: 95%;
-            cursor: col-resize;
-            background: transparent;
-            z-index: 11;
-        }
-        .resize-handle:hover {
-            background-color: var(--vscode-focusBorder);
-        }
-        .results-table td {
-            border: 1px solid var(--vscode-widget-border);
-            padding: 6px 8px;
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-            cursor: pointer;
-            user-select: none;
-            position: relative;
-            /* Prevent focus outline issues */
-            outline: none !important;
-        }
-        .results-table td:focus,
-        .results-table td:focus-visible {
-            outline: none !important;
-            box-shadow: none !important;
-        }
-        /* Prevent selection highlighting on shift+click */
-        .results-table,
-        .results-table * {
-            -webkit-user-select: none;
-            -moz-user-select: none;
-            -ms-user-select: none;
-            user-select: none;
-        }
-        .results-table td::-moz-focus-inner,
-        .detail-button-cell::-moz-focus-inner {
-            border: 0;
-            outline: none;
-        }
-        .results-table td:hover {
-            background-color: var(--vscode-list-hoverBackground);
-        }
-        .results-table td.selected {
-            background-color: var(--vscode-list-activeSelectionBackground) !important;
-            color: var(--vscode-list-activeSelectionForeground);
-        }
-        .results-table td[title] {
-            /* Ensure tooltips are properly displayed */
-            white-space: nowrap;
-        }
-        /* JSON content styling */
-        .results-table td .json-content {
-            font-family: var(--vscode-editor-font-family, 'Consolas', 'Monaco', monospace);
-            color: var(--vscode-textPreformat-foreground);
-        }
-        
-        /* Custom tooltip styling */
-        .custom-tooltip {
-            position: absolute;
-            background: var(--vscode-editorHoverWidget-background);
-            border: 1px solid var(--vscode-editorHoverWidget-border);
-            border-radius: 4px;
-            padding: 8px 12px;
-            font-size: 0.9em;
-            color: var(--vscode-editorHoverWidget-foreground);
-            z-index: 1000;
-            max-width: 400px;
-            max-height: 300px;
-            overflow: auto;
-            white-space: pre-wrap;
-            font-family: var(--vscode-editor-font-family, 'Consolas', 'Monaco', monospace);
-            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-            pointer-events: none;
-            opacity: 0;
-            transition: opacity 0.15s ease-in-out;
-        }
-        
-        .custom-tooltip.interactive {
-            pointer-events: auto;
-            user-select: text;
-            -webkit-user-select: text;
-            -moz-user-select: text;
-            -ms-user-select: text;
-        }
-        
-        .custom-tooltip.show {
-            opacity: 1;
-        }
-        
-        .custom-tooltip.json {
-            font-family: var(--vscode-editor-font-family, 'Consolas', 'Monaco', monospace);
-        }
-        
-        /* Custom context menu styling */
-        .custom-context-menu {
-            position: absolute;
-            background: var(--vscode-menu-background);
-            border: 1px solid var(--vscode-menu-border);
-            border-radius: 4px;
-            box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-            z-index: 2000;
-            min-width: 160px;
-            padding: 4px 0;
-            font-size: 0.9em;
-        }
-        
-        .context-menu-item {
-            padding: 8px 16px;
-            cursor: pointer;
-            color: var(--vscode-menu-foreground);
-            display: flex;
-            align-items: center;
-            gap: 8px;
-        }
-        
-        .context-menu-item:hover {
-            background: var(--vscode-menu-selectionBackground);
-            color: var(--vscode-menu-selectionForeground);
-        }
-        
-        .context-menu-item.disabled {
-            color: var(--vscode-disabledForeground);
-            cursor: default;
-        }
-        
-        .context-menu-item.disabled:hover {
-            background: transparent;
-            color: var(--vscode-disabledForeground);
-        }
-        
-        /* Loading indicator styling */
-        .loading-overlay {
-            position: absolute;
-            top: 0;
-            left: 0;
-            right: 0;
-            bottom: 0;
-            background: rgba(var(--vscode-editor-background-rgb, 30, 30, 30), 0.7);
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            z-index: 100;
-            backdrop-filter: blur(2px);
-        }
-        
-        .loading-content {
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-            gap: 16px;
-            color: var(--vscode-foreground);
-            font-size: 0.9em;
-            text-align: center;
-        }
-        
-        .loading-animation {
-            width: 128px;
-            height: 128px;
-            border-radius: 8px;
-            overflow: hidden;
-        }
-        
-        .loading-animation img {
-            width: 100%;
-            height: 100%;
-            object-fit: contain;
-        }
-        
-        .loading-message {
-            font-weight: 500;
-            max-width: 200px;
-        }
-        .results-table tr:nth-child(even) {
-            background-color: var(--vscode-editor-widget-background);
-        }
-        .no-results {
-            text-align: center;
-            padding: 40px;
-            color: var(--vscode-descriptionForeground);
-            font-style: italic;
-            position: absolute;
-            top: 50%;
-            left: 50%;
-            transform: translate(-50%, -50%);
-            width: 100%;
-        }
-        .error {
-            color: var(--vscode-errorForeground);
-            background-color: var(--vscode-inputValidation-errorBackground);
-            border: 1px solid var(--vscode-inputValidation-errorBorder);
-            padding: 15px;
-            border-radius: 4px;
-            margin: 15px;
-        }
-    </style>
-</head>
-<body>
-    <div class="main-content" id="mainContent">
-        <div class="content-wrapper" id="contentWrapper">
-            <div class="table-section" id="tableSection">
-                <div id="tableContainer" class="table-container">
-                    <div class="no-results">No results to display</div>
-                </div>
-            </div>
-            
-            <div id="resizeHandle" class="resize-handle" style="display: none;"></div>
-            
-            <div class="details-section" id="detailsSection" style="display: none;">
-                <div id="detailsPanel" class="details-panel">
-                    <div class="details-header">
-                        <div class="details-title">Row Details</div>
-                        <div class="details-navigation">
-                            <button id="detailsPrevBtn" class="nav-btn" onclick="navigateDetails(-1)" title="Previous row">
-                                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                                    <polyline points="18,15 12,9 6,15"></polyline>
-                                </svg>
-                            </button>
-                            <button id="detailsNextBtn" class="nav-btn" onclick="navigateDetails(1)" title="Next row">
-                                <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-                                    <polyline points="6,9 12,15 18,9"></polyline>
-                                </svg>
-                            </button>
-                            <button id="detailsCloseBtn" class="close-btn" onclick="closeDetails()" title="Close details">
-                                <span>×</span>
-                            </button>
-                        </div>
-                    </div>
-                    <div id="detailsContent" class="details-content">
-                        <!-- Details content will be populated by JavaScript -->
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    
-    <div class="footer">
-        <div id="resultsInfo" class="results-info">No results yet. Run a query to see results here.</div>
-        <button id="exportBtn" class="export-btn" onclick="exportToCsv()" style="display: none;">Export CSV</button>
-    </div>
-
-    <script>
-        const vscode = acquireVsCodeApi();
-        const loadingGifUri = '${loadingGifUri}';
-        let currentResults = null;
-        let sortState = { column: null, direction: null };
-
-        function escapeHtml(text) {
-            if (typeof text !== 'string') {
-                text = String(text);
-            }
-            const div = document.createElement('div');
-            div.textContent = text;
-            return div.innerHTML;
-        }
-
-        function formatCellValue(cell) {
-            if (cell === null || cell === undefined) {
-                return { displayValue: '<em style="color: var(--vscode-descriptionForeground);">null</em>', tooltipValue: 'null' };
-            }
-            
-            if (typeof cell === 'object') {
-                try {
-                    // Format JSON with proper indentation for tooltip
-                    const jsonString = JSON.stringify(cell, null, 2);
-                    
-                    // Create a compact display version for the cell
-                    const compactJson = JSON.stringify(cell);
-                    
-                    // Truncate display if too long
-                    const maxDisplayLength = 100;
-                    let displayValue = compactJson.length > maxDisplayLength 
-                        ? compactJson.substring(0, maxDisplayLength) + '...' 
-                        : compactJson;
-                    
-                    // Add JSON styling
-                    displayValue = '<span class="json-content">' + escapeHtml(displayValue) + '</span>';
-                    
-                    return {
-                        displayValue: displayValue,
-                        tooltipValue: jsonString
-                    };
-                } catch (error) {
-                    // If JSON.stringify fails, fall back to string conversion
-                    const stringValue = String(cell);
-                    return {
-                        displayValue: escapeHtml(stringValue),
-                        tooltipValue: stringValue
-                    };
-                }
-            }
-            
-            // For non-object values, use as-is
-            const stringValue = String(cell);
-            return {
-                displayValue: escapeHtml(stringValue),
-                tooltipValue: stringValue
-            };
-        }
-
-        function displayResults(result) {
-            currentResults = result;
-            
-            // Close details pane when new query results are displayed
-            if (selectedDetailRowIndices.length > 0) {
-                closeDetails();
-            }
-            
-            const tableContainer = document.getElementById('tableContainer');
-            const resultsInfo = document.getElementById('resultsInfo');
-            const exportBtn = document.getElementById('exportBtn');
-            
-            if (!result.columns || !result.data || result.data.length === 0) {
-                tableContainer.innerHTML = '<div class="no-results">No results found.</div>';
-                resultsInfo.textContent = 'No results';
-                exportBtn.style.display = 'none';
-                return;
-            }
-
-            const executionTimeText = result.executionTimeMs ? 
-                ' • ' + result.executionTimeMs + 'ms' : '';
-            resultsInfo.textContent = result.totalRecords + ' results' + executionTimeText + ' • ' + new Date(result.timestamp).toLocaleString();
-            exportBtn.style.display = 'block';
-            
-            let tableHtml = '<table class="results-table"><thead><tr>';
-            
-            // Add details button column header
-            tableHtml += '<th class="detail-button-cell select-all-header" style="width: 36px; cursor: pointer; text-align: center; padding: 0;" title="Click to select/deselect all rows" draggable="false" ondragstart="return false;" onclick="toggleSelectAllRows()">' +
-                '<svg viewBox="0 0 16 16" style="width: 18px; height: 18px; fill: var(--vscode-descriptionForeground); opacity: 0.7; vertical-align: middle; margin: 0; transition: all 0.2s ease;">' +
-                '<circle cx="6.5" cy="6.5" r="4" fill="none" stroke="currentColor" stroke-width="1.5"/>' +
-                '<path d="m9.5 9.5 4 4" stroke="currentColor" stroke-width="1.5" stroke-linecap="round"/>' +
-                '</svg>' +
-                '</th>';
-            
-            result.columns.forEach((col, index) => {
-                const sortClass = sortState.column === index ? 
-                    (sortState.direction === 'asc' ? 'sorted-asc' : 'sorted-desc') : '';
-                tableHtml += '<th class="' + sortClass + '" ' +
-                    'draggable="true" ' +
-                    'data-col-index="' + index + '" ' +
-                    'onclick="handleHeaderClick(event, ' + index + ')" ' +
-                    'ondragstart="handleDragStart(event, ' + index + ')"' +
-                    'ondragover="handleDragOver(event)"' +
-                    'ondrop="handleDrop(event, ' + index + ')"' +
-                    'ondragend="handleDragEnd(event)"' +
-                    'style="width: ' + (col.width || 'auto') + ';"' +
-                    'title="' + col.name + '">' +
-                    '<span class="header-text">' + col.name + '</span>' +
-                    '<div class="resize-handle" onmousedown="startResize(event, ' + index + ')"></div>' +
-                '</th>';
-            });
-            
-            tableHtml += '</tr></thead><tbody>';
-            
-            result.data.forEach((row, rowIndex) => {
-                tableHtml += '<tr>';
-                
-                // Add details button cell  
-                const isSelected = selectedDetailRowIndices.includes(rowIndex);
-                const circleIcon = isSelected ? 
-                    '<circle cx="8" cy="8" r="5" fill="currentColor"/>' :
-                    '<circle cx="8" cy="8" r="5" fill="none" stroke="currentColor" stroke-width="2"/>';
-                
-                tableHtml += '<td class="detail-button-cell">' +
-                    '<button class="detail-button' + (isSelected ? ' active' : '') + '" onclick="showRowDetails(' + rowIndex + ')" title="View row details">' +
-                    '<svg viewBox="0 0 16 16">' +
-                    circleIcon +
-                    '</svg>' +
-                    '</button>' +
-                '</td>';
-                
-                row.forEach((cell, cellIndex) => {
-                    const { displayValue, tooltipValue } = formatCellValue(cell);
-                    // Store tooltip data as data attribute - tooltipValue is already safe
-                    tableHtml += '<td data-tooltip="' + tooltipValue.replace(/"/g, '&quot;') + '" ' +
-                        'onclick="selectCell(this, ' + rowIndex + ', ' + cellIndex + ')" ' +
-                        'onmousedown="startCellDrag(event, this, ' + rowIndex + ', ' + cellIndex + ')"' +
-                        'onmouseenter="handleCellDragEnter(event, this, ' + rowIndex + ', ' + cellIndex + ')" ' +
-                        'onmouseleave="hideCustomTooltipDelayed()" ' +
-                        'data-row="' + rowIndex + '" ' +
-                        'data-col="' + cellIndex + '">' + displayValue + '</td>';
-                });
-                tableHtml += '</tr>';
-            });
-            
-            tableHtml += '</tbody></table>';
-            tableContainer.innerHTML = tableHtml;
-            
-            // Update detail button states after table regeneration
-            setTimeout(() => {
-                updateDetailButtonStates();
-            }, 0);
-            
-            // Add context menu event listener to the table
-            const table = tableContainer.querySelector('.results-table');
-            if (table) {
-                table.addEventListener('contextmenu', handleTableContextMenu);
-            }
-        }
-
-        // Loading indicator functionality
-        const loadingMessages = [
-            "Finding Azure treasures...",
-            "Sailing the cloudy seas...",
-            "Mapping uncharted resources…",
-            "Charting resource providers...",
-            "Plundering key vaults...",
-            "Investigating secrets...",
-            "Exploring landing zones...",
-            "Following effective routes...",
-            "Opening route maps...",
-            "Digging for hidden properties...",
-            "Navigating configuration drift..."
-        ];
-
-        function getRandomLoadingMessage() {
-            return loadingMessages[Math.floor(Math.random() * loadingMessages.length)];
-        }
-
-        function showLoadingIndicator() {
-            const contentWrapper = document.getElementById('contentWrapper');
-            if (!contentWrapper) return;
-            
-            // Remove any existing loading overlay
-            hideLoadingIndicator();
-            
-            const loadingOverlay = document.createElement('div');
-            loadingOverlay.className = 'loading-overlay';
-            loadingOverlay.id = 'loadingOverlay';
-            
-            const randomMessage = getRandomLoadingMessage();
-            
-            loadingOverlay.innerHTML = 
-                '<div class="loading-content">' +
-                    '<div class="loading-animation">' +
-                        '<img src="' + loadingGifUri + '" alt="Loading..." />' +
-                    '</div>' +
-                    '<div class="loading-message">' + randomMessage + '</div>' +
-                '</div>';
-            
-            contentWrapper.style.position = 'relative';
-            contentWrapper.appendChild(loadingOverlay);
-        }
-
-        function hideLoadingIndicator() {
-            const loadingOverlay = document.getElementById('loadingOverlay');
-            if (loadingOverlay) {
-                loadingOverlay.remove();
-            }
-        }
-
-        // Custom tooltip functionality
-        let customTooltip = null;
-        let tooltipTimeout = null;
-        let hideTooltipTimeout = null;
-
-        function createTooltip() {
-            if (!customTooltip) {
-                customTooltip = document.createElement('div');
-                customTooltip.className = 'custom-tooltip';
-                
-                // Add mouse events to tooltip for interactivity
-                customTooltip.addEventListener('mouseenter', function() {
-                    // Cancel any pending hide when mouse enters tooltip
-                    if (hideTooltipTimeout) {
-                        clearTimeout(hideTooltipTimeout);
-                        hideTooltipTimeout = null;
-                    }
-                });
-                
-                customTooltip.addEventListener('mouseleave', function(e) {
-                    // Don't hide tooltip if mouse is moving to the context menu
-                    // Check if the related target (where mouse is going) is the context menu
-                    if (contextMenuVisible && customContextMenu && 
-                        (customContextMenu.contains(e.relatedTarget) || e.relatedTarget === customContextMenu)) {
-                        return;
-                    }
-                    
-                    // Hide tooltip when mouse leaves tooltip
-                    hideCustomTooltipDelayed();
-                });
-                
-                // Custom context menu for tooltip with just Copy option
-                customTooltip.addEventListener('contextmenu', function(e) {
-                    e.preventDefault();
-                    e.stopPropagation();
-                    
-                    // Only show context menu if there's text selected in the tooltip
-                    const selection = window.getSelection();
-                    if (selection && selection.toString().length > 0) {
-                        showTooltipContextMenu(e);
-                    }
-                });
-                
-                document.body.appendChild(customTooltip);
-            }
-            return customTooltip;
-        }
-
-        function showCustomTooltip(event, element) {
-            // Don't show tooltip if context menu is visible
-            if (contextMenuVisible) {
-                return;
-            }
-            
-            const tooltip = createTooltip();
-            const tooltipText = element.getAttribute('data-tooltip');
-            
-            if (!tooltipText || tooltipText.trim() === '') {
-                // Don't show tooltip if content is empty
-                hideCustomTooltip();
-                return;
-            }
-            
-            // Clear any existing timeouts
-            if (tooltipTimeout) {
-                clearTimeout(tooltipTimeout);
-                tooltipTimeout = null;
-            }
-            if (hideTooltipTimeout) {
-                clearTimeout(hideTooltipTimeout);
-                hideTooltipTimeout = null;
-            }
-            
-            tooltip.textContent = tooltipText;
-            
-            // Add JSON class if content looks like JSON
-            if (tooltipText.trim().startsWith('{') || tooltipText.trim().startsWith('[')) {
-                tooltip.classList.add('json');
-            } else {
-                tooltip.classList.remove('json');
-            }
-            
-            // Position tooltip anchored to the cell, not cursor
-            positionTooltipToCell(element, tooltip);
-            
-            // Show tooltip immediately
-            tooltip.classList.add('show');
-            
-            // Check if tooltip has scrollable content (vertical or horizontal) and make it interactive
-            setTimeout(() => {
-                const hasVerticalScroll = tooltip.scrollHeight > tooltip.clientHeight;
-                const hasHorizontalScroll = tooltip.scrollWidth > tooltip.clientWidth;
-                
-                if (hasVerticalScroll || hasHorizontalScroll) {
-                    tooltip.classList.add('interactive');
-                } else {
-                    tooltip.classList.remove('interactive');
-                }
-            }, 0);
-        }
-
-        function positionTooltipToCell(cellElement, tooltip) {
-            // Get cell dimensions and position
-            const cellRect = cellElement.getBoundingClientRect();
-            const offset = 2; // Smaller offset for tighter positioning
-            
-            // Get viewport dimensions
-            const viewportWidth = window.innerWidth;
-            const viewportHeight = window.innerHeight;
-            
-            // Set initial position to measure tooltip dimensions
-            tooltip.style.left = '0px';
-            tooltip.style.top = '0px';
-            tooltip.style.visibility = 'hidden';
-            tooltip.style.display = 'block';
-            
-            // Get tooltip dimensions after it's rendered
-            const tooltipRect = tooltip.getBoundingClientRect();
-            const tooltipWidth = tooltipRect.width;
-            const tooltipHeight = tooltipRect.height;
-            
-            // Reset visibility
-            tooltip.style.visibility = 'visible';
-            
-            // Calculate primary anchor point: right 10% of cell horizontally, top of cell vertically
-            const rightAnchorX = cellRect.left + (cellRect.width * 0.90);
-            const anchorY = cellRect.top;
-            
-            // Try positioning tooltip's top-left corner at the right anchor point first
-            let left = rightAnchorX + offset;
-            let top = anchorY;
-            
-            // Check if tooltip would go off-screen to the right
-            if (left + tooltipWidth > viewportWidth - offset) {
-                // Position to the left using left 10% anchor point
-                const leftAnchorX = cellRect.left + (cellRect.width * 0.1);
-                left = leftAnchorX - tooltipWidth - offset;
-            }
-            
-            // Final fallback: ensure tooltip doesn't go off-screen to the left
-            if (left < offset) {
-                left = offset;
-            }
-            
-            // Adjust vertical position if tooltip would go off-screen
-            if (top < offset) {
-                top = offset;
-            } else if (top + tooltipHeight > viewportHeight - offset) {
-                top = viewportHeight - tooltipHeight - offset;
-            }
-            
-            tooltip.style.left = left + 'px';
-            tooltip.style.top = top + 'px';
-        }
-
-        function hideCustomTooltip() {
-            if (customTooltip) {
-                customTooltip.classList.remove('show');
-                customTooltip.classList.remove('interactive');
-                
-                // Remove tooltip after transition
-                tooltipTimeout = setTimeout(() => {
-                    if (customTooltip && !customTooltip.classList.contains('show')) {
-                        customTooltip.textContent = '';
-                    }
-                }, 150); // Match transition duration
-            }
-        }
-        
-        function hideCustomTooltipDelayed() {
-            // Don't hide tooltip if context menu is visible
-            if (contextMenuVisible) {
-                return;
-            }
-            
-            // Add a small delay to allow mouse to move from cell to tooltip
-            hideTooltipTimeout = setTimeout(() => {
-                hideCustomTooltip();
-            }, 100);
-        }
-
-        // Tooltip context menu functionality
-        let tooltipContextMenu = null;
-        let storedTooltipSelection = null; // Store the selection text when right-clicking
-        
-        function showTooltipContextMenu(event) {
-            // Store the current selection before any menu operations
-            const selection = window.getSelection();
-            storedTooltipSelection = selection && selection.toString().length > 0 ? selection.toString() : null;
-            
-            // Hide any existing context menus
-            hideContextMenu();
-            hideTooltipContextMenu();
-            
-            const tooltipMenu = createTooltipContextMenu();
-            positionContextMenu(event, tooltipMenu);
-            
-            // Set the context menu as visible
-            contextMenuVisible = true;
-            
-            // Add event listeners to hide menu when clicking outside or pressing Escape
-            setTimeout(() => {
-                document.addEventListener('click', hideTooltipContextMenu, { once: true });
-                document.addEventListener('keydown', (e) => {
-                    if (e.key === 'Escape') {
-                        hideTooltipContextMenu();
-                    }
-                }, { once: true });
-            }, 0);
-        }
-
-        function createTooltipContextMenu() {
-            // Remove existing tooltip menu if any
-            if (tooltipContextMenu) {
-                document.body.removeChild(tooltipContextMenu);
-            }
-            
-            tooltipContextMenu = document.createElement('div');
-            tooltipContextMenu.className = 'custom-context-menu';
-            
-            // Prevent clicks on the menu itself from closing it
-            tooltipContextMenu.addEventListener('click', (e) => {
-                e.stopPropagation();
-            });
-            
-            // Handle mouse leave from context menu
-            tooltipContextMenu.addEventListener('mouseleave', (e) => {
-                // If mouse is going back to tooltip, don't hide anything
-                if (customTooltip && customTooltip.contains(e.relatedTarget)) {
-                    return;
-                }
-                
-                // If mouse is going somewhere else, hide the tooltip after a short delay
-                setTimeout(() => {
-                    if (!contextMenuVisible) {
-                        hideCustomTooltipDelayed();
-                    }
-                }, 50);
-            });
-            
-            // Copy option for selected text in tooltip
-            const copyItem = document.createElement('div');
-            copyItem.className = 'context-menu-item';
-            copyItem.innerHTML = '<span>Copy</span>';
-            copyItem.addEventListener('click', () => {
-                copyTooltipSelection();
-                hideTooltipContextMenu();
-            });
-            
-            tooltipContextMenu.appendChild(copyItem);
-            document.body.appendChild(tooltipContextMenu);
-            
-            return tooltipContextMenu;
-        }
-
-        function hideTooltipContextMenu() {
-            if (tooltipContextMenu) {
-                document.body.removeChild(tooltipContextMenu);
-                tooltipContextMenu = null;
-            }
-            // Reset context menu visibility flag
-            contextMenuVisible = false;
-        }
-
-        function copyTooltipSelection() {
-            // Use the stored selection from when the context menu was opened
-            const selectedText = storedTooltipSelection;
-            
-            if (selectedText && selectedText.length > 0) {
-                // Copy to clipboard
-                navigator.clipboard.writeText(selectedText).then(() => {
-                    // Clear the stored selection
-                    storedTooltipSelection = null;
-                    // Hide the tooltip context menu since we've completed the action
-                    hideTooltipContextMenu();
-                }).catch(err => {
-                    console.error('Failed to copy text: ', err);
-                    // Fallback for older browsers
-                    const textArea = document.createElement('textarea');
-                    textArea.value = selectedText;
-                    document.body.appendChild(textArea);
-                    textArea.select();
-                    document.execCommand('copy');
-                    document.body.removeChild(textArea);
-                    
-                    // Clear stored selection and hide menu
-                    storedTooltipSelection = null;
-                    hideTooltipContextMenu();
-                });
-            }
-        }
-
-        // Cell selection functionality
-        let selectedCells = new Set();
-        let isSelecting = false;
-        let selectionStart = null;
-        let isDragging = false;
-        let dragStartCell = null;
-        let dragCurrentCell = null;
-
-        function selectCell(cellElement, row, col) {
-            // Don't handle click if we just finished a drag operation
-            if (isDragging) {
-                return;
-            }
-            
-            if (event.ctrlKey || event.metaKey) {
-                // Ctrl/Cmd click for multi-select
-                toggleCellSelection(cellElement, row, col);
-            } else if (event.shiftKey && selectionStart) {
-                // Shift click for range selection
-                selectRange(selectionStart.row, selectionStart.col, row, col);
-            } else {
-                // Regular click - clear previous selection and select this cell
-                clearSelection();
-                toggleCellSelection(cellElement, row, col);
-                selectionStart = { row, col };
-            }
-        }
-
-        function toggleCellSelection(cellElement, row, col) {
-            const cellKey = row + '-' + col;
-            if (selectedCells.has(cellKey)) {
-                selectedCells.delete(cellKey);
-                cellElement.classList.remove('selected');
-            } else {
-                selectedCells.add(cellKey);
-                cellElement.classList.add('selected');
-            }
-        }
-
-        function selectRange(startRow, startCol, endRow, endCol) {
-            clearSelection();
-            
-            const minRow = Math.min(startRow, endRow);
-            const maxRow = Math.max(startRow, endRow);
-            const minCol = Math.min(startCol, endCol);
-            const maxCol = Math.max(startCol, endCol);
-            
-            for (let r = minRow; r <= maxRow; r++) {
-                for (let c = minCol; c <= maxCol; c++) {
-                    const cellElement = document.querySelector('td[data-row="' + r + '"][data-col="' + c + '"]');
-                    if (cellElement) {
-                        const cellKey = r + '-' + c;
-                        selectedCells.add(cellKey);
-                        cellElement.classList.add('selected');
-                    }
-                }
-            }
-        }
-
-        function clearSelection() {
-            selectedCells.clear();
-            document.querySelectorAll('.results-table td.selected').forEach(cell => {
-                cell.classList.remove('selected');
-            });
-        }
-
-        // Drag selection functionality
-        function startCellDrag(event, cellElement, row, col) {
-            // Don't start drag if it's a right-click (we want context menu)
-            if (event.button === 2) {
-                return;
-            }
-            
-            // Prevent text selection during drag
-            event.preventDefault();
-            document.body.style.userSelect = 'none';
-            
-            isDragging = true;
-            dragStartCell = { row, col, element: cellElement };
-            dragCurrentCell = { row, col, element: cellElement };
-            
-            // Clear previous selection if not using Ctrl/Cmd
-            if (!event.ctrlKey && !event.metaKey) {
-                clearSelection();
-            }
-            
-            // Select the starting cell
-            toggleCellSelection(cellElement, row, col);
-            selectionStart = { row, col };
-            
-            // Add document-level mouse up listener to end drag
-            document.addEventListener('mouseup', endCellDrag, { once: true });
-            
-            // Prevent the default click handler from firing
-            event.stopPropagation();
-        }
-
-        function handleCellDragEnter(event, cellElement, row, col) {
-            // Only handle tooltip if we're not dragging
-            if (!isDragging) {
-                showCustomTooltip(event, cellElement);
-                return;
-            }
-            
-            // Handle drag selection
-            if (isDragging && dragStartCell) {
-                dragCurrentCell = { row, col, element: cellElement };
-                
-                // Clear current selection (but keep Ctrl/Cmd behavior)
-                if (!event.ctrlKey && !event.metaKey) {
-                    clearSelection();
-                }
-                
-                // Select range from start to current cell
-                selectRange(dragStartCell.row, dragStartCell.col, row, col);
-            }
-        }
-
-        function endCellDrag(event) {
-            if (isDragging) {
-                isDragging = false;
-                dragStartCell = null;
-                dragCurrentCell = null;
-                
-                // Re-enable text selection
-                document.body.style.userSelect = '';
-            }
-        }
-
-        // Custom context menu functionality
-        let customContextMenu = null;
-        let rightClickedCell = null; // Track the cell that was right-clicked
-        let contextMenuVisible = false; // Track if context menu is showing
-
-        function handleTableContextMenu(event) {
-            event.preventDefault();
-            
-            // Find the cell that was right-clicked
-            const target = event.target.closest('td');
-            
-            if (!target) return;
-            
-            // Get row and column from data attributes
-            const rowAttr = target.getAttribute('data-row');
-            const colAttr = target.getAttribute('data-col');
-            
-            if (rowAttr === null || colAttr === null) return;
-            
-            const row = parseInt(rowAttr);
-            const col = parseInt(colAttr);
-            
-            if (isNaN(row) || isNaN(col)) return;
-            
-            // Store the right-clicked cell info
-            rightClickedCell = {
-                element: target,
-                row: row,
-                col: col
-            };
-            
-            // Smart selection logic
-            const cellKey = row + '-' + col;
-            const isRightClickedCellSelected = selectedCells.has(cellKey);
-            
-            if (selectedCells.size <= 1) {
-                // If we have 0 or 1 cells selected, select the right-clicked cell
-                selectCell(target, row, col);
-            } else if (selectedCells.size > 1 && !isRightClickedCellSelected) {
-                // If we have multiple cells selected but right-clicked outside the selection,
-                // clear selection and select just the right-clicked cell
-                clearSelection();
-                selectCell(target, row, col);
-            }
-            // If multiple cells selected and right-clicked on one of them, keep current selection
-            
-            // Hide tooltip and its context menu when showing cell context menu
-            hideCustomTooltip();
-            hideTooltipContextMenu();
-            
-            // Hide any existing cell context menu
-            if (customContextMenu) {
-                document.body.removeChild(customContextMenu);
-                customContextMenu = null;
-            }
-            
-            // Create context menu
-            const contextMenu = createContextMenu();
-            
-            // Position and show context menu
-            positionContextMenu(event, contextMenu);
-            
-            // Set flag that context menu is visible
-            contextMenuVisible = true;
-            
-            // Add click outside listener to close menu
-            setTimeout(() => {
-                document.addEventListener('click', hideContextMenu, { once: true });
-                document.addEventListener('keydown', handleContextMenuKeydown, { once: true });
-            }, 0);
-        }
-        
-        function handleContextMenuKeydown(event) {
-            if (event.key === 'Escape') {
-                hideContextMenu();
-            }
-        }
-
-        function createContextMenu() {
-            if (customContextMenu) {
-                document.body.removeChild(customContextMenu);
-            }
-            
-            customContextMenu = document.createElement('div');
-            customContextMenu.className = 'custom-context-menu';
-            
-            // Prevent clicks on the menu itself from closing it
-            customContextMenu.addEventListener('click', (e) => {
-                e.stopPropagation();
-            });
-            
-            // Check if we have a valid right-clicked cell
-            let hasValidRightClick = false;
-            let hasJsonCell = false;
-            let rightClickedCellKey = null;
-            let cellIsNull = false;
-            
-            if (rightClickedCell && currentResults && currentResults.data) {
-                const { row, col } = rightClickedCell;
-                
-                if (currentResults.data[row] && currentResults.data[row][col] !== undefined) {
-                    const cellValue = currentResults.data[row][col];
-                    cellIsNull = cellValue === null || cellValue === undefined;
-                    hasValidRightClick = !cellIsNull; // Only valid if not null
-                    rightClickedCellKey = row + '-' + col;
-                    if (typeof cellValue === 'object' && cellValue !== null) {
-                        hasJsonCell = true;
-                    }
-                }
-            }
-            
-            // Check if right-clicked cell is part of current selection
-            const hasSelection = selectedCells.size > 0;
-            const rightClickedSelectedCell = hasSelection && rightClickedCellKey && selectedCells.has(rightClickedCellKey);
-            
-            // Check if any selected cells contain JSON (for formatted copy option)
-            let hasSelectedJsonCells = false;
-            if (hasSelection) {
-                selectedCells.forEach(cellKey => {
-                    const [row, col] = cellKey.split('-').map(Number);
-                    if (currentResults && currentResults.data && currentResults.data[row] && currentResults.data[row][col] !== undefined) {
-                        const cellValue = currentResults.data[row][col];
-                        if (typeof cellValue === 'object' && cellValue !== null) {
-                            hasSelectedJsonCells = true;
-                        }
-                    }
-                });
-            }
-            
-            // Copy single cell option (disabled if cell is null OR if multiple cells are selected)
-            let copyItem = null;
-            const hasMultipleSelected = rightClickedSelectedCell && hasSelection && selectedCells.size > 1;
-            
-            // Only show single cell copy if we don't have multiple cells selected
-            if (!hasMultipleSelected) {
-                copyItem = document.createElement('div');
-                copyItem.className = 'context-menu-item ' + (hasValidRightClick ? '' : 'disabled');
-                copyItem.innerHTML = '<span>Copy</span>';
-                copyItem.addEventListener('click', () => {
-                    if (hasValidRightClick) {
-                        copyRightClickedCell();
-                        hideContextMenu();
-                    }
-                });
-            }
-            
-            // Copy selection option (only if right-clicked on a selected cell and there are multiple selected)
-            let copySelectionItem = null;
-            if (rightClickedSelectedCell && hasSelection && selectedCells.size > 1) {
-                copySelectionItem = document.createElement('div');
-                copySelectionItem.className = 'context-menu-item';
-                copySelectionItem.innerHTML = '<span>Copy Selection (' + selectedCells.size + ' cells)</span>';
-                copySelectionItem.addEventListener('click', () => {
-                    copySelectedCells();
-                    hideContextMenu();
-                });
-            }
-            
-            // Copy with headers option (only if right-clicked on a selected cell with multiple selections)
-            let copyWithHeadersItem = null;
-            if (rightClickedSelectedCell && hasSelection && selectedCells.size > 1) {
-                copyWithHeadersItem = document.createElement('div');
-                copyWithHeadersItem.className = 'context-menu-item';
-                copyWithHeadersItem.innerHTML = '<span>Copy Selection with Headers</span>';
-                copyWithHeadersItem.addEventListener('click', () => {
-                    copySelectedCellsWithHeaders();
-                    hideContextMenu();
-                });
-            }
-            
-            // Copy formatted option for single cell (only show if right-clicked cell contains JSON AND not in multi-selection context)
-            let copyFormattedItem = null;
-            if (hasJsonCell && !(rightClickedSelectedCell && selectedCells.size > 1)) {
-                copyFormattedItem = document.createElement('div');
-                copyFormattedItem.className = 'context-menu-item';
-                copyFormattedItem.innerHTML = '<span>Copy Formatted</span>';
-                copyFormattedItem.addEventListener('click', () => {
-                    copyRightClickedCellFormatted();
-                    hideContextMenu();
-                });
-            }
-            
-            // Copy selection formatted option - REMOVED per user request
-            // Don't show formatted option for multi-cell selections
-            let copySelectionFormattedItem = null;
-            
-            // Add all menu items
-            if (copyItem) {
-                customContextMenu.appendChild(copyItem);
-            }
-            if (copySelectionItem) {
-                customContextMenu.appendChild(copySelectionItem);
-            }
-            if (copyWithHeadersItem) {
-                customContextMenu.appendChild(copyWithHeadersItem);
-            }
-            if (copyFormattedItem) {
-                customContextMenu.appendChild(copyFormattedItem);
-            }
-            if (copySelectionFormattedItem) {
-                customContextMenu.appendChild(copySelectionFormattedItem);
-            }
-            
-            document.body.appendChild(customContextMenu);
-            return customContextMenu;
-        }
-
-        function positionContextMenu(event, menu) {
-            const mouseX = event.clientX;
-            const mouseY = event.clientY;
-            
-            // Get viewport dimensions
-            const viewportWidth = window.innerWidth;
-            const viewportHeight = window.innerHeight;
-            
-            // Set initial position to measure menu dimensions
-            menu.style.left = '0px';
-            menu.style.top = '0px';
-            menu.style.visibility = 'hidden';
-            menu.style.display = 'block';
-            
-            const menuRect = menu.getBoundingClientRect();
-            const menuWidth = menuRect.width;
-            const menuHeight = menuRect.height;
-            
-            // Calculate position
-            let left = mouseX;
-            let top = mouseY;
-            
-            // Adjust if menu would go off-screen
-            if (left + menuWidth > viewportWidth) {
-                left = mouseX - menuWidth;
-            }
-            
-            if (top + menuHeight > viewportHeight) {
-                top = mouseY - menuHeight;
-            }
-            
-            // Ensure menu doesn't go off-screen
-            left = Math.max(0, left);
-            top = Math.max(0, top);
-            
-            menu.style.left = left + 'px';
-            menu.style.top = top + 'px';
-            menu.style.visibility = 'visible';
-        }
-
-        function hideContextMenu() {
-            if (customContextMenu) {
-                document.body.removeChild(customContextMenu);
-                customContextMenu = null;
-            }
-            // Reset the right-clicked cell when menu is actually hidden
-            rightClickedCell = null;
-            // Reset context menu visibility flag
-            contextMenuVisible = false;
-            // Remove any lingering event listeners
-            document.removeEventListener('keydown', handleContextMenuKeydown);
-        }
-
-        function copySelectedCellsWithHeaders() {
-            if (selectedCells.size === 0) return;
-            
-            // Get all selected cells with their positions
-            const cellData = [];
-            selectedCells.forEach(cellKey => {
-                const [row, col] = cellKey.split('-').map(Number);
-                if (currentResults && currentResults.data[row] && currentResults.data[row][col] !== undefined) {
-                    cellData.push({
-                        row,
-                        col,
-                        value: currentResults.data[row][col] || ''
-                    });
-                }
-            });
-            
-            if (cellData.length === 0) return;
-            
-            // Sort by row then by column
-            cellData.sort((a, b) => a.row - b.row || a.col - b.col);
-            
-            // Get unique columns and their headers
-            const uniqueCols = [...new Set(cellData.map(cell => cell.col))].sort((a, b) => a - b);
-            const headers = uniqueCols.map(colIndex => {
-                return currentResults && currentResults.columns[colIndex] 
-                    ? currentResults.columns[colIndex].name 
-                    : 'Column ' + colIndex;
-            });
-            
-            // Group by rows
-            const rowGroups = {};
-            cellData.forEach(cell => {
-                if (!rowGroups[cell.row]) {
-                    rowGroups[cell.row] = {};
-                }
-                // Format the cell value for copying
-                let copyValue = cell.value;
-                if (typeof copyValue === 'object' && copyValue !== null) {
-                    copyValue = JSON.stringify(copyValue);
-                }
-                rowGroups[cell.row][cell.col] = copyValue || '';
-            });
-            
-            // Create header row
-            const headerRow = headers.join('\\t');
-            
-            // Create data rows
-            const rows = Object.keys(rowGroups).sort((a, b) => Number(a) - Number(b));
-            const dataRows = rows.map(rowIndex => {
-                const row = rowGroups[rowIndex];
-                return uniqueCols.map(colIndex => String(row[colIndex] || '')).join('\\t');
-            });
-            
-            // Combine header and data
-            const copyText = [headerRow, ...dataRows].join('\\n');
-            
-            // Copy to clipboard
-            navigator.clipboard.writeText(copyText).then(() => {
-                showCopyFeedback('with headers');
-            }).catch(err => {
-                console.error('Failed to copy to clipboard:', err);
-                fallbackCopyTextToClipboard(copyText);
-            });
-        }
-
-        function copySelectedCellsFormatted() {
-            if (selectedCells.size === 0) return;
-            
-            // Get all selected cells with their positions
-            const cellData = [];
-            selectedCells.forEach(cellKey => {
-                const [row, col] = cellKey.split('-').map(Number);
-                if (currentResults && currentResults.data[row] && currentResults.data[row][col] !== undefined) {
-                    cellData.push({
-                        row,
-                        col,
-                        value: currentResults.data[row][col] || ''
-                    });
-                }
-            });
-            
-            if (cellData.length === 0) return;
-            
-            // Sort by row then by column
-            cellData.sort((a, b) => a.row - b.row || a.col - b.col);
-            
-            // Format cells for copying - use formatted JSON for objects, regular text for others
-            const formattedCells = cellData.map(cell => {
-                let copyValue = cell.value;
-                if (typeof copyValue === 'object' && copyValue !== null) {
-                    // Use formatted JSON (like in tooltip) for objects
-                    try {
-                        copyValue = JSON.stringify(copyValue, null, 2);
-                    } catch (error) {
-                        copyValue = String(copyValue);
-                    }
-                } else {
-                    copyValue = String(copyValue || '');
-                }
-                return copyValue;
-            });
-            
-            // Join with newlines for multi-cell selection, or just return single cell
-            const copyText = formattedCells.join('\\n\\n');
-            
-            // Copy to clipboard
-            navigator.clipboard.writeText(copyText).then(() => {
-                showCopyFeedback('formatted');
-            }).catch(err => {
-                console.error('Failed to copy to clipboard:', err);
-                fallbackCopyTextToClipboard(copyText);
-            });
-        }
-
-        function copyRightClickedCell() {
-            if (!rightClickedCell || !currentResults || !currentResults.data ||
-                isNaN(rightClickedCell.row) || isNaN(rightClickedCell.col) ||
-                rightClickedCell.row < 0 || rightClickedCell.col < 0 ||
-                rightClickedCell.row >= currentResults.data.length ||
-                rightClickedCell.col >= (currentResults.data[rightClickedCell.row] || []).length) {
-                return;
-            }
-            
-            const cellValue = currentResults.data[rightClickedCell.row][rightClickedCell.col];
-            let copyValue = cellValue;
-            
-            // Format the cell value for copying
-            if (typeof copyValue === 'object' && copyValue !== null) {
-                copyValue = JSON.stringify(copyValue);
-            } else {
-                copyValue = String(copyValue || '');
-            }
-            
-            // Copy to clipboard
-            navigator.clipboard.writeText(copyValue).then(() => {
-                showCopyFeedback('');
-            }).catch(err => {
-                console.error('Failed to copy to clipboard:', err);
-                fallbackCopyTextToClipboard(copyValue);
-            });
-        }
-
-        function copyRightClickedCellFormatted() {
-            if (!rightClickedCell || !currentResults || !currentResults.data ||
-                isNaN(rightClickedCell.row) || isNaN(rightClickedCell.col) ||
-                rightClickedCell.row < 0 || rightClickedCell.col < 0 ||
-                rightClickedCell.row >= currentResults.data.length ||
-                rightClickedCell.col >= (currentResults.data[rightClickedCell.row] || []).length) {
-                return;
-            }
-            
-            const cellValue = currentResults.data[rightClickedCell.row][rightClickedCell.col];
-            let copyValue = cellValue;
-            
-            // Format the cell value for copying
-            if (typeof copyValue === 'object' && copyValue !== null) {
-                // Use formatted JSON (like in tooltip) for objects
-                try {
-                    copyValue = JSON.stringify(copyValue, null, 2);
-                } catch (error) {
-                    copyValue = String(copyValue);
-                }
-            } else {
-                copyValue = String(copyValue || '');
-            }
-            
-            // Copy to clipboard
-            navigator.clipboard.writeText(copyValue).then(() => {
-                showCopyFeedback('formatted');
-            }).catch(err => {
-                console.error('Failed to copy to clipboard:', err);
-                fallbackCopyTextToClipboard(copyValue);
-            });
-        }
-
-        function sortTable(columnIndex) {
-            if (!currentResults || !currentResults.data) return;
-            
-            let direction = 'asc';
-            if (sortState.column === columnIndex && sortState.direction === 'asc') {
-                direction = 'desc';
-            }
-            
-            sortState = { column: columnIndex, direction };
-            
-            const sortedData = [...currentResults.data].sort((a, b) => {
-                const aVal = a[columnIndex];
-                const bVal = b[columnIndex];
-                
-                if (aVal == null && bVal == null) return 0;
-                if (aVal == null) return direction === 'asc' ? -1 : 1;
-                if (bVal == null) return direction === 'asc' ? 1 : -1;
-                
-                const aNum = Number(aVal);
-                const bNum = Number(bVal);
-                if (!isNaN(aNum) && !isNaN(bNum)) {
-                    return direction === 'asc' ? aNum - bNum : bNum - aNum;
-                }
-                
-                const aStr = String(aVal).toLowerCase();
-                const bStr = String(bVal).toLowerCase();
-                
-                if (direction === 'asc') {
-                    return aStr < bStr ? -1 : aStr > bStr ? 1 : 0;
-                } else {
-                    return aStr > bStr ? -1 : aStr < bStr ? 1 : 0;
-                }
-            });
-            
-            const sortedResult = { ...currentResults, data: sortedData };
-            displayResults(sortedResult);
-            
-            // Update details panel if it's open
-            updateDetailsAfterSort();
-        }
-
-        function exportToCsv() {
-            if (currentResults) {
-                vscode.postMessage({ 
-                    type: 'exportCsv', 
-                    payload: { 
-                        data: currentResults, 
-                        filename: 'barge-results-' + new Date().toISOString().replace(/[:.]/g, '-') + '.csv'
-                    } 
-                });
-            }
-        }
-
-        function selectRange(startRow, startCol, endRow, endCol) {
-            clearSelection();
-            
-            const minRow = Math.min(startRow, endRow);
-            const maxRow = Math.max(startRow, endRow);
-            const minCol = Math.min(startCol, endCol);
-            const maxCol = Math.max(startCol, endCol);
-            
-            for (let r = minRow; r <= maxRow; r++) {
-                for (let c = minCol; c <= maxCol; c++) {
-                    const cellElement = document.querySelector('td[data-row="' + r + '"][data-col="' + c + '"]');
-                    if (cellElement) {
-                        const cellKey = r + '-' + c;
-                        selectedCells.add(cellKey);
-                        cellElement.classList.add('selected');
-                    }
-                }
-            }
-        }
-
-        function clearSelection() {
-            selectedCells.clear();
-            document.querySelectorAll('.results-table td.selected').forEach(cell => {
-                cell.classList.remove('selected');
-            });
-        }
-
-        // Drag selection functionality
-        function startCellDrag(event, cellElement, row, col) {
-            // Don't start drag if it's a right-click (we want context menu)
-            if (event.button === 2) {
-                return;
-            }
-            
-            // Prevent text selection during drag
-            event.preventDefault();
-            document.body.style.userSelect = 'none';
-            
-            isDragging = true;
-            dragStartCell = { row, col, element: cellElement };
-            dragCurrentCell = { row, col, element: cellElement };
-            
-            // Clear previous selection if not using Ctrl/Cmd
-            if (!event.ctrlKey && !event.metaKey) {
-                clearSelection();
-            }
-            
-            // Select the starting cell
-            toggleCellSelection(cellElement, row, col);
-            selectionStart = { row, col };
-            
-            // Add document-level mouse up listener to end drag
-            document.addEventListener('mouseup', endCellDrag, { once: true });
-            
-            // Prevent the default click handler from firing
-            event.stopPropagation();
-        }
-
-        function handleCellDragEnter(event, cellElement, row, col) {
-            // Only handle tooltip if we're not dragging
-            if (!isDragging) {
-                showCustomTooltip(event, cellElement);
-                return;
-            }
-            
-            // Handle drag selection
-            if (isDragging && dragStartCell) {
-                dragCurrentCell = { row, col, element: cellElement };
-                
-                // Clear current selection (but keep Ctrl/Cmd behavior)
-                if (!event.ctrlKey && !event.metaKey) {
-                    clearSelection();
-                }
-                
-                // Select range from start to current cell
-                selectRange(dragStartCell.row, dragStartCell.col, row, col);
-            }
-        }
-
-        function endCellDrag(event) {
-            if (isDragging) {
-                isDragging = false;
-                dragStartCell = null;
-                dragCurrentCell = null;
-                
-                // Re-enable text selection
-                document.body.style.userSelect = '';
-            }
-        }
-
-        // Add keyboard support for cell selection
-        document.addEventListener('keydown', function(event) {
-            if (event.key === 'Escape') {
-                // Close details panel if open, otherwise clear selection
-                if (selectedDetailRowIndices.length > 0) {
-                    closeDetails();
-                } else {
-                    clearSelection();
-                    hideContextMenu(); // Also hide context menu on Escape
-                }
-            } else if (selectedDetailRowIndices.length === 1 && (event.key === 'ArrowUp' || event.key === 'ArrowDown')) {
-                // Navigate details panel with arrow keys when exactly one row is selected
-                event.preventDefault();
-                const direction = event.key === 'ArrowUp' ? -1 : 1;
-                navigateDetails(direction);
-            } else if ((event.ctrlKey || event.metaKey) && event.key === 'a') {
-                // Ctrl/Cmd+A to select all cells
-                event.preventDefault();
-                selectAllCells();
-            } else if ((event.ctrlKey || event.metaKey) && event.key === 'c') {
-                // Ctrl/Cmd+C to copy selected cells
-                event.preventDefault();
-                copySelectedCells();
-            }
-        });
-
-        function copySelectedCells() {
-            if (selectedCells.size === 0) return;
-            
-            // Get all selected cells with their positions
-            const cellData = [];
-            selectedCells.forEach(cellKey => {
-                const [row, col] = cellKey.split('-').map(Number);
-                const cellElement = document.querySelector('td[data-row="' + row + '"][data-col="' + col + '"]');
-                if (cellElement && currentResults && currentResults.data[row] && currentResults.data[row][col] !== undefined) {
-                    cellData.push({
-                        row,
-                        col,
-                        value: currentResults.data[row][col] || ''
-                    });
-                }
-            });
-            
-            if (cellData.length === 0) return;
-            
-            // Sort by row then by column
-            cellData.sort((a, b) => a.row - b.row || a.col - b.col);
-            
-            // Group by rows and create tab-separated text
-            const rowGroups = {};
-            cellData.forEach(cell => {
-                if (!rowGroups[cell.row]) {
-                    rowGroups[cell.row] = {};
-                }
-                // Format the cell value for copying (use raw value for objects, formatted for display)
-                let copyValue = cell.value;
-                if (typeof copyValue === 'object' && copyValue !== null) {
-                    copyValue = JSON.stringify(copyValue);
-                }
-                rowGroups[cell.row][cell.col] = copyValue || '';
-            });
-            
-            // Convert to tab-separated format
-            const rows = Object.keys(rowGroups).sort((a, b) => Number(a) - Number(b));
-            const copyText = rows.map(rowIndex => {
-                const row = rowGroups[rowIndex];
-                const cols = Object.keys(row).sort((a, b) => Number(a) - Number(b));
-                
-                // If it's a single row selection, just join the values with tabs
-                if (rows.length === 1) {
-                    return cols.map(colIndex => String(row[colIndex] || '')).join('\\t');
-                }
-                
-                // For multiple rows, we need to handle gaps in column selection
-                const minCol = Math.min(...cols.map(Number));
-                const maxCol = Math.max(...cols.map(Number));
-                const rowData = [];
-                
-                for (let i = minCol; i <= maxCol; i++) {
-                    rowData.push(String(row[i] || ''));
-                }
-                
-                return rowData.join('\\t');
-            }).join('\\n');
-            
-            // Copy to clipboard
-            navigator.clipboard.writeText(copyText).then(() => {
-                // Visual feedback that copy was successful
-                showCopyFeedback();
-            }).catch(err => {
-                console.error('Failed to copy to clipboard:', err);
-                // Fallback for older browsers
-                fallbackCopyTextToClipboard(copyText);
-            });
-        }
-        
-        function fallbackCopyTextToClipboard(text) {
-            const textArea = document.createElement('textarea');
-            textArea.value = text;
-            textArea.style.position = 'fixed';
-            textArea.style.left = '-999999px';
-            textArea.style.top = '-999999px';
-            document.body.appendChild(textArea);
-            textArea.focus();
-            textArea.select();
-            
-            try {
-                document.execCommand('copy');
-                showCopyFeedback();
-            } catch (err) {
-                console.error('Fallback copy failed:', err);
-            }
-            
-            document.body.removeChild(textArea);
-        }
-        
-        function showCopyFeedback(type = '') {
-            // Create temporary visual feedback
-            const feedback = document.createElement('div');
-            const cellCount = selectedCells.size;
-            const typeText = type ? ' ' + type : '';
-            feedback.textContent = 'Copied ' + cellCount + ' cell' + (cellCount === 1 ? '' : 's') + typeText;
-            feedback.style.position = 'fixed';
-            feedback.style.top = '10px';
-            feedback.style.right = '10px';
-            feedback.style.background = 'var(--vscode-notifications-background)';
-            feedback.style.color = 'var(--vscode-notifications-foreground)';
-            feedback.style.padding = '8px 12px';
-            feedback.style.borderRadius = '4px';
-            feedback.style.border = '1px solid var(--vscode-notifications-border)';
-            feedback.style.fontSize = '0.9em';
-            feedback.style.zIndex = '1000';
-            feedback.style.boxShadow = '0 2px 8px rgba(0,0,0,0.1)';
-            
-            document.body.appendChild(feedback);
-            
-            // Remove after 2 seconds
-            setTimeout(() => {
-                if (feedback.parentNode) {
-                    document.body.removeChild(feedback);
-                }
-            }, 2000);
-        }
-
-        function selectAllCells() {
-            clearSelection();
-            document.querySelectorAll('.results-table td[data-row]').forEach(cell => {
-                const row = cell.getAttribute('data-row');
-                const col = cell.getAttribute('data-col');
-                const cellKey = row + '-' + col;
-                selectedCells.add(cellKey);
-                cell.classList.add('selected');
-            });
-        }
-
-        // Column resizing functionality
-        let isResizing = false;
-        let currentResizeColumn = null;
-        let startX = 0;
-        let startWidth = 0;
-
-        function startResize(event, columnIndex) {
-            event.stopPropagation();
-            event.preventDefault();
-            
-            isResizing = true;
-            currentResizeColumn = columnIndex;
-            startX = event.clientX;
-            
-            const th = event.target.closest('th');
-            startWidth = th.offsetWidth;
-            
-            document.addEventListener('mousemove', handleResize);
-            document.addEventListener('mouseup', stopResize);
-            
-            // Prevent text selection during resize
-            document.body.style.userSelect = 'none';
-        }
-
-        function handleResize(event) {
-            if (!isResizing || currentResizeColumn === null) return;
-            
-            const diff = event.clientX - startX;
-            const newWidth = Math.max(60, startWidth + diff); // Minimum width of 60px
-            
-            const th = document.querySelector('th[data-col-index="' + currentResizeColumn + '"]');
-            if (th) {
-                th.style.width = newWidth + 'px';
-                
-                // Store the width in our data
-                if (currentResults && currentResults.columns[currentResizeColumn]) {
-                    currentResults.columns[currentResizeColumn].width = newWidth + 'px';
-                }
-            }
-        }
-
-        function stopResize() {
-            isResizing = false;
-            currentResizeColumn = null;
-            document.removeEventListener('mousemove', handleResize);
-            document.removeEventListener('mouseup', stopResize);
-            document.body.style.userSelect = '';
-        }
-
-        // Column drag and drop functionality
-        let draggedColumn = null;
-
-        function handleHeaderClick(event, columnIndex) {
-            // Only sort if we're not clicking on the resize handle
-            if (!event.target.classList.contains('resize-handle')) {
-                sortTable(columnIndex);
-            }
-        }
-
-        function handleDragStart(event, columnIndex) {
-            // Don't start drag if we're on the resize handle
-            if (event.target.classList.contains('resize-handle')) {
-                event.preventDefault();
-                return;
-            }
-            
-            draggedColumn = columnIndex;
-            event.target.classList.add('dragging');
-            
-            // Set drag data
-            event.dataTransfer.effectAllowed = 'move';
-            event.dataTransfer.setData('text/html', event.target.outerHTML);
-        }
-
-        function handleDragOver(event) {
-            if (draggedColumn === null) return;
-            
-            event.preventDefault();
-            event.dataTransfer.dropEffect = 'move';
-            
-            const th = event.target.closest('th');
-            if (th && th !== event.target.closest('table').querySelector('.dragging')) {
-                // Remove previous drag-over classes
-                document.querySelectorAll('.drag-over').forEach(el => el.classList.remove('drag-over'));
-                th.classList.add('drag-over');
-            }
-        }
-
-        function handleDrop(event, targetColumnIndex) {
-            event.preventDefault();
-            
-            if (draggedColumn === null || draggedColumn === targetColumnIndex) {
-                return;
-            }
-            
-            // Reorder the columns in our data
-            reorderColumn(draggedColumn, targetColumnIndex);
-            
-            // Clear drag states
-            document.querySelectorAll('.drag-over').forEach(el => el.classList.remove('drag-over'));
-        }
-
-        function handleDragEnd(event) {
-            event.target.classList.remove('dragging');
-            document.querySelectorAll('.drag-over').forEach(el => el.classList.remove('drag-over'));
-            draggedColumn = null;
-        }
-
-        function reorderColumn(fromIndex, toIndex) {
-            if (!currentResults || !currentResults.columns || !currentResults.data) return;
-            
-            // Reorder columns metadata
-            const columns = [...currentResults.columns];
-            const [movedColumn] = columns.splice(fromIndex, 1);
-            columns.splice(toIndex, 0, movedColumn);
-            
-            // Reorder data
-            const data = currentResults.data.map(row => {
-                const newRow = [...row];
-                const [movedCell] = newRow.splice(fromIndex, 1);
-                newRow.splice(toIndex, 0, movedCell);
-                return newRow;
-            });
-            
-            // Update our current results
-            currentResults = {
-                ...currentResults,
-                columns,
-                data
-            };
-            
-            // Update sort state if needed
-            if (sortState.column !== null) {
-                if (sortState.column === fromIndex) {
-                    sortState.column = toIndex;
-                } else if (fromIndex < toIndex && sortState.column > fromIndex && sortState.column <= toIndex) {
-                    sortState.column--;
-                } else if (fromIndex > toIndex && sortState.column >= toIndex && sortState.column < fromIndex) {
-                    sortState.column++;
-                }
-            }
-            
-            // Clear selection since column indices have changed
-            clearSelection();
-            
-            // Re-render the table
-            displayResults(currentResults);
-        }
-
-        function displayError(error) {
-            const tableContainer = document.getElementById('tableContainer');
-            const resultsInfo = document.getElementById('resultsInfo');
-            const exportBtn = document.getElementById('exportBtn');
-            
-            tableContainer.innerHTML = '<div class="error">Error: ' + error + '</div>';
-            resultsInfo.textContent = 'Query failed';
-            exportBtn.style.display = 'none';
-        }
-
-        // Details panel functionality
-        let currentDetailRowIndex = -1;
-        let selectedDetailRowIndices = []; // Array to track multiple selected rows
-        let currentDetailRowData = null;
-        
-        function updateDetailButtonStates() {
-            const buttons = document.querySelectorAll('.detail-button');
-            buttons.forEach((button, index) => {
-                const isActive = selectedDetailRowIndices.includes(index);
-                const svg = button.querySelector('svg');
-                if (svg) {
-                    const circle = svg.querySelector('circle');
-                    if (circle) {
-                        if (isActive) {
-                            circle.setAttribute('fill', 'currentColor');
-                            circle.removeAttribute('stroke');
-                            circle.removeAttribute('stroke-width');
-                            circle.setAttribute('r', '5');
-                            button.classList.add('active');
-                        } else {
-                            circle.setAttribute('fill', 'none');
-                            circle.setAttribute('stroke', 'currentColor');
-                            circle.setAttribute('stroke-width', '2');
-                            circle.setAttribute('r', '5');
-                            button.classList.remove('active');
-                        }
-                    }
-                }
-            });
-            
-            // Update magnifying glass header state
-            const headerMagnifyingGlass = document.querySelector('.select-all-header svg circle');
-            if (headerMagnifyingGlass) {
-                const isComparisonActive = selectedDetailRowIndices.length > 1;
-                if (isComparisonActive) {
-                    // Add blue fill while keeping the grey outline
-                    headerMagnifyingGlass.setAttribute('fill', 'var(--vscode-focusBorder)');
-                    headerMagnifyingGlass.setAttribute('stroke', 'currentColor');
-                    headerMagnifyingGlass.setAttribute('stroke-width', '1.5');
-                } else {
-                    // Reset to outline only
-                    headerMagnifyingGlass.setAttribute('fill', 'none');
-                    headerMagnifyingGlass.setAttribute('stroke', 'currentColor');
-                    headerMagnifyingGlass.setAttribute('stroke-width', '1.5');
-                }
-            }
-        }
-        
-        function selectEntireRow(rowIndex) {
-            if (!currentResults || !currentResults.columns || rowIndex < 0) {
-                return;
-            }
-            
-            // Don't clear selection here anymore - it's handled by the caller
-            
-            // Select all cells in the row (excluding the details button column which is at index -1)
-            for (let colIndex = 0; colIndex < currentResults.columns.length; colIndex++) {
-                const cellElement = document.querySelector('td[data-row="' + rowIndex + '"][data-col="' + colIndex + '"]');
-                if (cellElement) {
-                    const cellKey = rowIndex + '-' + colIndex;
-                    selectedCells.add(cellKey);
-                    cellElement.classList.add('selected');
-                }
-            }
-        }
-        
-        function toggleSelectAllRows() {
-            if (!currentResults || !currentResults.data) {
-                return;
-            }
-            
-            // Check if all rows are currently selected
-            const allSelected = currentResults.data.length > 0 && 
-                               selectedDetailRowIndices.length === currentResults.data.length &&
-                               selectedDetailRowIndices.every((index, i) => index === i);
-            
-            if (allSelected) {
-                // Deselect all rows and close details pane
-                selectedDetailRowIndices = [];
-                // Clear all cell selections
-                clearSelection();
-                updateDetailButtonStates();
-                closeDetails();
-            } else {
-                // Select all rows (or remaining rows)
-                selectedDetailRowIndices = [];
-                for (let i = 0; i < currentResults.data.length; i++) {
-                    selectedDetailRowIndices.push(i);
-                }
-                
-                // Update the detail buttons visual state
-                updateDetailButtonStates();
-                
-                // Select all cells in all selected rows (clear previous selection first)
-                clearSelection();
-                selectedDetailRowIndices.forEach(idx => selectEntireRow(idx));
-                
-                // Show comparison view for all rows
-                if (selectedDetailRowIndices.length > 1) {
-                    // Open details pane properly with correct sizing
-                    const detailsSection = document.getElementById('detailsSection');
-                    const resizeHandle = document.getElementById('resizeHandle');
-                    const tableSection = document.getElementById('tableSection');
-                    
-                    // Detect and set proper layout
-                    detectAndSetLayout();
-                    
-                    // Show the details section and resize handle
-                    if (detailsSection) detailsSection.style.display = 'flex';
-                    if (resizeHandle) resizeHandle.style.display = 'block';
-                    
-                    // Set initial flex values for 2:1 ratio (table:details)
-                    if (tableSection && detailsSection) {
-                        tableSection.style.flex = '2 1 0';
-                        detailsSection.style.flex = '1 1 0';
-                    }
-                    
-                    // Set initial resize handle position
-                    setTimeout(() => {
-                        const resizeHandle = document.getElementById('resizeHandle');
-                        if (resizeHandle) {
-                            if (currentLayout === 'horizontal') {
-                                resizeHandle.style.left = '66.666%';
-                                resizeHandle.style.top = '2.5%'; // Account for 95% height with 2.5% margin
-                            } else {
-                                resizeHandle.style.top = 'calc(66.666% - 3px)';
-                                resizeHandle.style.left = '2.5%'; // Account for 95% width with 2.5% margin
-                            }
-                            initializeResizing();
-                        }
-                    }, 0);
-                    
-                    const detailsContent = document.getElementById('detailsContent');
-                    if (detailsContent) {
-                        detailsContent.innerHTML = generateComparisonView(selectedDetailRowIndices);
-                        detailsContent.classList.add('comparison-view');
-                        initializeComparisonColumnResizing();
-                    }
-                } else if (selectedDetailRowIndices.length === 1) {
-                    // Show single row details
-                    showRowDetails(selectedDetailRowIndices[0]);
-                }
-            }
-        }
-        
-        function showRowDetails(rowIndex) {
-            if (!currentResults || !currentResults.data || rowIndex < 0 || rowIndex >= currentResults.data.length) {
-                return;
-            }
-            
-            // Toggle selection for multiple row comparison
-            const existingIndex = selectedDetailRowIndices.indexOf(rowIndex);
-            if (existingIndex >= 0) {
-                // Remove from selection
-                selectedDetailRowIndices.splice(existingIndex, 1);
-                if (selectedDetailRowIndices.length === 0) {
-                    closeDetails();
-                    return;
-                }
-            } else {
-                // Add to selection
-                selectedDetailRowIndices.push(rowIndex);
-            }
-            
-            // For backward compatibility, keep currentDetailRowIndex as the first selected
-            currentDetailRowIndex = selectedDetailRowIndices[0] || -1;
-            currentDetailRowData = selectedDetailRowIndices.length > 0 ? currentResults.data[selectedDetailRowIndices[0]] : null;
-            
-            // Update button states to show active state
-            updateDetailButtonStates();
-            
-            // Select all selected rows (clear previous selection first)
-            clearSelection();
-            selectedDetailRowIndices.forEach(idx => selectEntireRow(idx));
-            
-            const contentWrapper = document.getElementById('contentWrapper');
-            const detailsSection = document.getElementById('detailsSection');
-            const tableSection = document.getElementById('tableSection');
-            const resizeHandle = document.getElementById('resizeHandle');
-            const detailsContent = document.getElementById('detailsContent');
-            const detailsTitle = document.querySelector('.details-title');
-            
-            // Detect and set layout
-            detectAndSetLayout();
-            
-            // Show the details section and resize handle
-            detailsSection.style.display = 'flex';
-            resizeHandle.style.display = 'block';
-            
-            // Set initial flex values for 2:1 ratio (table:details)
-            if (tableSection && detailsSection) {
-                tableSection.style.flex = '2 1 0';
-                detailsSection.style.flex = '1 1 0';
-            }
-            
-            // Set initial resize handle position
-            setTimeout(() => {
-                const resizeHandle = document.getElementById('resizeHandle');
-                if (resizeHandle) {
-                    if (currentLayout === 'horizontal') {
-                        resizeHandle.style.left = '66.666%';
-                        resizeHandle.style.top = '2.5%'; // Account for 95% height with 2.5% margin
-                    } else {
-                        resizeHandle.style.top = 'calc(66.666% - 3px)';
-                        resizeHandle.style.left = '2.5%'; // Account for 95% width with 2.5% margin
-                    }
-                }
-            }, 0);
-            
-            // Initialize resizing if not already done
-            initializeResizing();
-            
-            // Update title to show selected rows
-            if (selectedDetailRowIndices.length === 1) {
-                const rowNumber = selectedDetailRowIndices[0] + 1;
-                const totalRows = currentResults.data.length;
-                detailsTitle.textContent = 'Row ' + rowNumber + ' of ' + totalRows;
-            } else {
-                detailsTitle.textContent = 'Comparing ' + selectedDetailRowIndices.length + ' rows';
-            }
-            
-            // Update navigation buttons
-            updateDetailsNavigation();
-            
-            // Generate and display the comparison or single row details
-            if (selectedDetailRowIndices.length === 1) {
-                const rowData = generateRowObject(selectedDetailRowIndices[0]);
-                const jsonHtml = formatAsJsonViewer(rowData);
-                detailsContent.innerHTML = jsonHtml;
-                // Remove comparison view class for single row
-                detailsContent.classList.remove('comparison-view');
-            } else {
-                const comparisonHtml = generateComparisonView(selectedDetailRowIndices);
-                detailsContent.innerHTML = comparisonHtml;
-                // Add comparison view class for reduced padding
-                detailsContent.classList.add('comparison-view');
-                // Initialize column resizing for comparison table
-                initializeComparisonColumnResizing();
-            }
-        }
-        
-        function closeDetails() {
-            const detailsSection = document.getElementById('detailsSection');
-            const resizeHandle = document.getElementById('resizeHandle');
-            const tableSection = document.getElementById('tableSection');
-            const detailsContent = document.getElementById('detailsContent');
-            
-            detailsSection.style.display = 'none';
-            resizeHandle.style.display = 'none';
-            
-            // Reset flex values
-            if (tableSection) tableSection.style.flex = '';
-            if (detailsSection) detailsSection.style.flex = '';
-            
-            // Remove comparison view class
-            if (detailsContent) {
-                detailsContent.classList.remove('comparison-view');
-            }
-            
-            currentDetailRowIndex = -1;
-            selectedDetailRowIndices = []; // Clear multiple selection
-            currentDetailRowData = null;
-            
-            // Update button states to remove active state
-            updateDetailButtonStates();
-        }
-        
-        function updateDetailsAfterSort() {
-            // If details panel is open, find the new index of the tracked row data
-            if (currentDetailRowData && currentResults && currentResults.data) {
-                const newIndex = currentResults.data.findIndex(row => {
-                    // Compare objects by all their properties
-                    if (!row || typeof row !== 'object') return false;
-                    
-                    const keys1 = Object.keys(currentDetailRowData);
-                    const keys2 = Object.keys(row);
-                    
-                    if (keys1.length !== keys2.length) return false;
-                    
-                    return keys1.every(key => currentDetailRowData[key] === row[key]);
-                });
-                
-                if (newIndex !== -1) {
-                    currentDetailRowIndex = newIndex;
-                    
-                    // Update the details panel title
-                    const detailsTitle = document.querySelector('.details-title');
-                    if (detailsTitle) {
-                        const rowNumber = newIndex + 1;
-                        const totalRows = currentResults.data.length;
-                        detailsTitle.textContent = 'Row ' + rowNumber + ' of ' + totalRows;
-                    }
-                    
-                    // Update navigation buttons
-                    updateDetailsNavigation();
-                    
-                    // Select the row at its new position
-                    selectEntireRow(newIndex);
-                }
-            }
-        }
-        
-        function navigateDetails(direction) {
-            if (!currentResults || !currentResults.data || selectedDetailRowIndices.length !== 1) {
-                return; // Only allow navigation for single selection
-            }
-            
-            const currentIndex = selectedDetailRowIndices[0];
-            const newIndex = currentIndex + direction;
-            
-            if (newIndex >= 0 && newIndex < currentResults.data.length) {
-                // Update selection
-                selectedDetailRowIndices = [newIndex];
-                currentDetailRowIndex = newIndex;
-                currentDetailRowData = currentResults.data[newIndex];
-                
-                // Update UI without going through full showRowDetails
-                updateDetailButtonStates();
-                selectEntireRow(newIndex);
-                
-                // Update content
-                const detailsTitle = document.querySelector('.details-title');
-                const detailsContent = document.getElementById('detailsContent');
-                
-                const rowNumber = newIndex + 1;
-                const totalRows = currentResults.data.length;
-                detailsTitle.textContent = 'Row ' + rowNumber + ' of ' + totalRows;
-                
-                const rowData = generateRowObject(newIndex);
-                const jsonHtml = formatAsJsonViewer(rowData);
-                detailsContent.innerHTML = jsonHtml;
-                
-                updateDetailsNavigation();
-            }
-        }
-        
-        function updateDetailsNavigation() {
-            const prevBtn = document.getElementById('detailsPrevBtn');
-            const nextBtn = document.getElementById('detailsNextBtn');
-            
-            if (!currentResults || !currentResults.data || selectedDetailRowIndices.length !== 1) {
-                prevBtn.disabled = true;
-                nextBtn.disabled = true;
-                return;
-            }
-            
-            const currentIndex = selectedDetailRowIndices[0];
-            prevBtn.disabled = currentIndex <= 0;
-            nextBtn.disabled = currentIndex >= currentResults.data.length - 1;
-        }
-        
-        function generateRowObject(rowIndex) {
-            if (!currentResults || !currentResults.columns || !currentResults.data || 
-                rowIndex < 0 || rowIndex >= currentResults.data.length) {
-                return {};
-            }
-            
-            const row = currentResults.data[rowIndex];
-            const rowObject = {};
-            
-            currentResults.columns.forEach((column, colIndex) => {
-                if (colIndex < row.length) {
-                    rowObject[column.name] = row[colIndex];
-                }
-            });
-            
-            return rowObject;
-        }
-        
-        function generateComparisonView(rowIndices) {
-            if (!currentResults || !currentResults.data || rowIndices.length === 0) {
-                return '<div class="comparison-viewer">No data available</div>';
-            }
-            
-            // Get all unique property names across selected rows
-            const allProperties = new Set();
-            const rowObjects = rowIndices.map(idx => generateRowObject(idx));
-            
-            rowObjects.forEach(row => {
-                Object.keys(row).forEach(key => allProperties.add(key));
-            });
-            
-            let html = '<div class="comparison-viewer">';
-            html += '<table class="comparison-table">';
-            html += '<thead><tr>';
-            html += '<th class="property-name-header">Property<div class="column-resizer"></div></th>';
-            
-            // Add column headers for each selected row
-            rowIndices.forEach((rowIdx, colIdx) => {
-                const isLast = colIdx === rowIndices.length - 1;
-                html += '<th class="row-header">Row ' + (rowIdx + 1);
-                if (!isLast) {
-                    html += '<div class="column-resizer"></div>';
-                }
-                html += '</th>';
-            });
-            html += '</tr></thead><tbody>';
-            
-            // Create a row for each property
-            Array.from(allProperties).sort().forEach(property => {
-                html += '<tr class="property-row">';
-                
-                // Get values for this property across all selected rows
-                const values = rowObjects.map(row => row[property]);
-                const uniqueValues = [...new Set(values.map(v => JSON.stringify(v)))];
-                const allMatch = uniqueValues.length === 1;
-                
-                // Add property name with match indicator
-                const tooltipText = allMatch ? "Values match" : "Values differ";
-                html += '<td class="property-name-with-indicator" title="' + tooltipText + '">';
-                html += '<span class="property-name-text">' + escapeHtml(property) + '</span>';
-                html += '<span class="match-indicator-inline">';
-                if (allMatch) {
-                    html += '<svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="#22c55e" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">';
-                    html += '<polyline points="20,6 9,17 4,12"></polyline>';
-                    html += '</svg>';
-                } else {
-                    html += '<svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="#ef4444" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">';
-                    html += '<line x1="18" y1="6" x2="6" y2="18"></line>';
-                    html += '<line x1="6" y1="6" x2="18" y2="18"></line>';
-                    html += '</svg>';
-                }
-                html += '</span>';
-                html += '</td>';
-                
-                // Add value columns
-                values.forEach(value => {
-                    html += '<td class="property-value">';
-                    if (value === null || value === undefined) {
-                        html += '<span class="null-value">null</span>';
-                    } else if (typeof value === 'object') {
-                        html += '<pre class="object-value">' + escapeHtml(JSON.stringify(value, null, 2)) + '</pre>';
-                    } else {
-                        html += '<span class="scalar-value">' + escapeHtml(String(value)) + '</span>';
-                    }
-                    html += '</td>';
-                });
-                
-                html += '</tr>';
-            });
-            
-            html += '</tbody></table></div>';
-            return html;
-        }
-        
-        function initializeComparisonColumnResizing() {
-            const table = document.querySelector('.comparison-table');
-            if (!table) return;
-            
-            const resizers = table.querySelectorAll('.column-resizer');
-            let isResizing = false;
-            let currentResizer = null;
-            let startX = 0;
-            let startWidth = 0;
-            
-            resizers.forEach(resizer => {
-                resizer.addEventListener('mousedown', (e) => {
-                    isResizing = true;
-                    currentResizer = resizer;
-                    startX = e.clientX;
-                    
-                    const th = resizer.parentElement;
-                    startWidth = th.offsetWidth;
-                    
-                    // Prevent text selection during resize
-                    e.preventDefault();
-                    document.body.style.userSelect = 'none';
-                    document.body.style.cursor = 'col-resize';
-                });
-            });
-            
-            document.addEventListener('mousemove', (e) => {
-                if (!isResizing || !currentResizer) return;
-                
-                const th = currentResizer.parentElement;
-                const diff = e.clientX - startX;
-                const newWidth = Math.max(50, startWidth + diff); // Minimum width of 50px
-                
-                th.style.width = newWidth + 'px';
-                th.style.minWidth = newWidth + 'px';
-                th.style.maxWidth = newWidth + 'px';
-            });
-            
-            document.addEventListener('mouseup', () => {
-                if (isResizing) {
-                    isResizing = false;
-                    currentResizer = null;
-                    document.body.style.userSelect = '';
-                    document.body.style.cursor = '';
-                }
-            });
-        }
-        
-        function formatAsJsonViewer(obj) {
-            if (!obj || typeof obj !== 'object') {
-                return '<div class="json-viewer">No data available</div>';
-            }
-            
-            let html = '<div class="json-viewer">';
-            
-            Object.keys(obj).forEach(key => {
-                const value = obj[key];
-                html += '<div class="json-property">';
-                html += '<div class="json-key">' + escapeHtml(key) + '</div>';
-                
-                if (value === null || value === undefined) {
-                    html += '<div class="json-value null">null</div>';
-                } else if (typeof value === 'string') {
-                    // Remove quotes for cleaner display, since we're in boxes now
-                    html += '<div class="json-value string">' + escapeHtml(value) + '</div>';
-                } else if (typeof value === 'number') {
-                    html += '<div class="json-value number">' + value + '</div>';
-                } else if (typeof value === 'boolean') {
-                    html += '<div class="json-value boolean">' + value + '</div>';
-                } else if (typeof value === 'object') {
-                    try {
-                        const jsonString = JSON.stringify(value, null, 2);
-                        html += '<pre class="json-value object">' + escapeHtml(jsonString) + '</pre>';
-                    } catch (error) {
-                        html += '<div class="json-value object">' + escapeHtml(String(value)) + '</div>';
-                    }
-                } else {
-                    html += '<div class="json-value">' + escapeHtml(String(value)) + '</div>';
-                }
-                
-                html += '</div>';
-            });
-            
-            html += '</div>';
-            return html;
-        }
-
-        window.addEventListener('message', event => {
-            const message = event.data;
-            switch (message.type) {
-                case 'queryStart':
-                    showLoadingIndicator();
-                    break;
-                case 'queryResult':
-                    hideLoadingIndicator();
-                    if (message.payload.success) {
-                        displayResults(message.payload.data);
-                    } else {
-                        displayError(message.payload.error);
-                    }
-                    break;
-            }
-        });
-
-        // Layout and resizing functionality
-        let isResizingPanel = false;
-        let resizeInitialized = false;
-        let currentLayout = 'horizontal'; // 'horizontal' or 'vertical'
-
-        function detectAndSetLayout() {
-            const contentWrapper = document.getElementById('contentWrapper');
-            if (!contentWrapper) return;
-            
-            const rect = contentWrapper.getBoundingClientRect();
-            const aspectRatio = rect.width / rect.height;
-            
-            // Use horizontal layout if width is significantly larger than height
-            const newLayout = aspectRatio > 1.2 ? 'horizontal' : 'vertical';
-            
-            if (newLayout !== currentLayout) {
-                currentLayout = newLayout;
-                updateLayoutClasses();
-            }
-        }
-
-        function updateLayoutClasses() {
-            const contentWrapper = document.getElementById('contentWrapper');
-            const resizeHandle = document.getElementById('resizeHandle');
-            if (!contentWrapper || !resizeHandle) return;
-            
-            contentWrapper.classList.remove('layout-horizontal', 'layout-vertical');
-            contentWrapper.classList.add('layout-' + currentLayout);
-            
-            // Reset any custom flex values when switching layouts
-            const tableSection = document.getElementById('tableSection');
-            const detailsSection = document.getElementById('detailsSection');
-            if (tableSection) tableSection.style.flex = '2 1 0';
-            if (detailsSection) detailsSection.style.flex = '1 1 0';
-            
-            // Position resize handle at 2:1 ratio
-            if (currentLayout === 'horizontal') {
-                resizeHandle.style.left = '66.666%';
-                resizeHandle.style.top = '2.5%'; // Account for 95% height with 2.5% margin
-            } else {
-                resizeHandle.style.top = 'calc(66.666% - 3px)'; // Account for handle height
-                resizeHandle.style.left = '2.5%'; // Account for 95% width with 2.5% margin
-            }
-        }
-
-        function initializeResizing() {
-            if (resizeInitialized) return;
-            resizeInitialized = true;
-            
-            const resizeHandle = document.getElementById('resizeHandle');
-            if (!resizeHandle) return;
-
-            resizeHandle.addEventListener('mousedown', startResizing);
-            
-            // Add window resize listener for layout detection
-            window.addEventListener('resize', detectAndSetLayout);
-        }
-
-        function startResizing(event) {
-            event.preventDefault();
-            isResizingPanel = true;
-            
-            const resizeHandle = document.getElementById('resizeHandle');
-            if (resizeHandle) {
-                resizeHandle.classList.add('dragging');
-            }
-            
-            const cursor = currentLayout === 'horizontal' ? 'ew-resize' : 'ns-resize';
-            document.body.style.cursor = cursor;
-            document.body.style.userSelect = 'none';
-            
-            document.addEventListener('mousemove', handleResizing);
-            document.addEventListener('mouseup', stopResizing);
-        }
-
-        function handleResizing(event) {
-            if (!isResizingPanel) return;
-            
-            const contentWrapper = document.getElementById('contentWrapper');
-            const tableSection = document.getElementById('tableSection');
-            const detailsSection = document.getElementById('detailsSection');
-            const resizeHandle = document.getElementById('resizeHandle');
-            
-            if (!contentWrapper || !tableSection || !detailsSection || !resizeHandle) return;
-            
-            const wrapperRect = contentWrapper.getBoundingClientRect();
-            
-            if (currentLayout === 'horizontal') {
-                // Horizontal layout - vertical divider
-                const mouseX = event.clientX;
-                const relativeX = mouseX - wrapperRect.left;
-                const percentage = Math.max(20, Math.min(80, (relativeX / wrapperRect.width) * 100));
-                
-                const tablePercentage = percentage;
-                const detailsPercentage = 100 - percentage;
-                
-                tableSection.style.flex = tablePercentage + ' 1 0';
-                detailsSection.style.flex = detailsPercentage + ' 1 0';
-                
-                // Update resize handle position
-                resizeHandle.style.left = percentage + '%';
-                resizeHandle.style.top = '2.5%'; // Account for 95% height with 2.5% margin
-            } else {
-                // Vertical layout - horizontal divider
-                const mouseY = event.clientY;
-                const relativeY = mouseY - wrapperRect.top;
-                const percentage = Math.max(20, Math.min(80, (relativeY / wrapperRect.height) * 100));
-                
-                const tablePercentage = percentage;
-                const detailsPercentage = 100 - percentage;
-                
-                tableSection.style.flex = tablePercentage + ' 1 0';
-                detailsSection.style.flex = detailsPercentage + ' 1 0';
-                
-                // Update resize handle position
-                resizeHandle.style.top = 'calc(' + percentage + '% - 3px)'; // Account for handle height
-                resizeHandle.style.left = '2.5%'; // Account for 95% width with 2.5% margin
-            }
-        }
-
-        function stopResizing() {
-            isResizingPanel = false;
-            document.body.style.cursor = '';
-            document.body.style.userSelect = '';
-            
-            const resizeHandle = document.getElementById('resizeHandle');
-            if (resizeHandle) {
-                resizeHandle.classList.remove('dragging');
-            }
-            
-            document.removeEventListener('mousemove', handleResizing);
-            document.removeEventListener('mouseup', stopResizing);
-        }
-        
-    </script>
-</body>
-</html>`;
+        const loadingGifUri = this._panel.webview.asWebviewUri(
+            vscode.Uri.joinPath(this._extensionUri, 'media', 'webview', 'loadingbarge.gif')
+        );
+
+        // Replace placeholders
+        html = html.replace(/{{NONCE}}/g, nonce);
+        html = html.replace(/{{WEBVIEW_URI}}/g, webviewUri.toString());
+        html = html.replace(/{{LOADING_GIF_URI}}/g, loadingGifUri.toString());
+
+        return html;
     }
 }
